test(tui): split mid-size TUI test blocks by topic
Apply the ADR 0028 `#[path = "..."] mod tests;` pattern to the six
remaining TUI files whose inline test blocks pushed the total past
the watch threshold. Split each into a sibling `<name>_tests/`
directory grouped by which prod function or behavior the tests pin.

- `tree.rs` (1620 → 646) → `tree_tests/{build_tree_structure,
  symbol_badges, file_size_warnings, skipped_files,
  test_files_and_overlap}`
- `order.rs` (1458 → 637) → `order_tests/{rank_directories,
  cycle_partners, cycle_edges, order_tree, scc_helpers}`
- `detail.rs` (1359 → 445) → `detail_tests/{build_detail,
  symbol_mentions, build_dir_detail, build_file_detail}`
- `diff_shape.rs` (1123 → 321) → `diff_shape_tests/{
  build_diff_pane_content, hunk_start_lines, section_start_line}`
- `row_view.rs` (1097 → 405) → `row_view_tests/{entry_row_line,
  relative_labels, file_size_badges, chg_ref_badges}`
- `ui/detail_pane.rs` (1168 → 326) → `ui/detail_pane_tests/{
  content_by_row_kind, scroll_behavior, wrap_reachability,
  size_warning}`

Shared fixtures move into each `<name>_tests/mod.rs` as
`pub(super) fn`. No production body changed. Test count unchanged
(524 rinkaku_tui tests still pass).
