+++ rinkaku-tui/src/detail.rs
@@ -441,919 +441,5 @@ fn collect_symbol_ids(node: &TreeNode, ids: &mut Vec<String>) {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use pretty_assertions::assert_eq;
-    use rinkaku_core::diff::LineRange;
-    use rinkaku_core::extract::ExtractedSymbol;
-    use rinkaku_core::graph::{Edge, Hotspot, Node, SymbolGraph};
-    use rinkaku_core::render::FileReport;
-
-    fn symbol(id: &str, name: &str) -> ExtractedSymbol {
-        ExtractedSymbol {
-            id: id.to_string(),
-            name: name.to_string(),
-            kind: SymbolKind::Function,
-            signature: format!("fn {name}()"),
-            range: LineRange { start: 1, end: 1 },
-            container: None,
-            referenced_names: vec![],
-            dependencies: vec![],
-            omitted_dependency_matches: 0,
-            is_test: false,
-            classification: None,
-            previous_signature: None,
-        }
-    }
-
-    fn node(id: &str, path: &str, name: &str) -> Node {
-        Node {
-            id: id.to_string(),
-            path: path.to_string(),
-            name: name.to_string(),
-        }
-    }
-
-    fn empty_report() -> Report {
-        Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![],
-                edges: vec![],
-                roots: vec![],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        }
-    }
-
-    #[test]
-    fn should_return_none_when_symbol_id_is_not_found() {
-        let report = empty_report();
-
-        let actual = build_detail(&report, "missing::id");
-
-        assert_eq!(None, actual);
-    }
-
-    #[test]
-    fn should_build_current_signature_when_classification_is_not_signature_changed() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![ExtractedSymbol {
-                    classification: Some(Classification::BodyOnly),
-                    ..symbol("lib.rs::foo", "foo")
-                }],
-            }],
-            ..empty_report()
-        };
-
-        let expected = DetailView {
-            id: "lib.rs::foo".to_string(),
-            name: "foo".to_string(),
-            kind: SymbolKind::Function,
-            path: "lib.rs".to_string(),
-            container: None,
-            signature: SignatureView::Current("fn foo()".to_string()),
-            classification: Some(Classification::BodyOnly),
-            used_by: vec![],
-            callees: vec![],
-            callers: vec![],
-        };
-        let actual = build_detail(&report, "lib.rs::foo");
-
-        assert_eq!(Some(expected), actual);
-    }
-
-    #[test]
-    fn should_build_changed_signature_when_classification_is_signature_changed() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![ExtractedSymbol {
-                    classification: Some(Classification::SignatureChanged),
-                    previous_signature: Some("fn foo(a: i32)".to_string()),
-                    signature: "fn foo(a: i32, b: i32)".to_string(),
-                    ..symbol("lib.rs::foo", "foo")
-                }],
-            }],
-            ..empty_report()
-        };
-
-        let actual = build_detail(&report, "lib.rs::foo").expect("symbol found");
-
-        let expected_signature = SignatureView::Changed {
-            previous: "fn foo(a: i32)".to_string(),
-            current: "fn foo(a: i32, b: i32)".to_string(),
-        };
-        assert_eq!(expected_signature, actual.signature);
-    }
-
-    #[test]
-    fn should_fall_back_to_current_signature_when_previous_signature_is_missing() {
-        // Defensive: SignatureChanged without a previous_signature (should
-        // not happen per `classify_symbols`'s contract) still renders
-        // something sane rather than panicking.
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![ExtractedSymbol {
-                    classification: Some(Classification::SignatureChanged),
-                    previous_signature: None,
-                    ..symbol("lib.rs::foo", "foo")
-                }],
-            }],
-            ..empty_report()
-        };
-
-        let actual = build_detail(&report, "lib.rs::foo").expect("symbol found");
-
-        assert_eq!(
-            SignatureView::Current("fn foo()".to_string()),
-            actual.signature
-        );
-    }
-
-    #[test]
-    fn should_list_single_caller_as_used_by_when_fan_in_is_one() {
-        // fan-in of exactly 1 is below Hotspot's >= 2 threshold, so
-        // `report.hotspots` has nothing for "callee" — used_by must still
-        // show the one caller by reading `graph.edges` directly.
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![
-                    symbol("lib.rs::caller", "caller"),
-                    symbol("lib.rs::callee", "callee"),
-                ],
-            }],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("lib.rs::caller", "lib.rs", "caller"),
-                    node("lib.rs::callee", "lib.rs", "callee"),
-                ],
-                edges: vec![Edge {
-                    from: "lib.rs::caller".to_string(),
-                    to: "lib.rs::callee".to_string(),
-                    is_cycle: false,
-                }],
-                roots: vec!["lib.rs::caller".to_string()],
-            },
-            hotspots: vec![],
-            ..empty_report()
-        };
-
-        let actual = build_detail(&report, "lib.rs::callee").expect("symbol found");
-
-        let expected_used_by = vec![SymbolMention {
-            id: "lib.rs::caller".to_string(),
-            name: "caller".to_string(),
-            path: "lib.rs".to_string(),
-        }];
-        assert_eq!(expected_used_by, actual.used_by);
-    }
-
-    #[test]
-    fn should_list_every_referrer_as_used_by_when_symbol_is_a_hotspot() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![
-                    symbol("lib.rs::a", "a"),
-                    symbol("lib.rs::b", "b"),
-                    symbol("lib.rs::shared", "shared"),
-                ],
-            }],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("lib.rs::a", "lib.rs", "a"),
-                    node("lib.rs::b", "lib.rs", "b"),
-                    node("lib.rs::shared", "lib.rs", "shared"),
-                ],
-                edges: vec![
-                    Edge {
-                        from: "lib.rs::a".to_string(),
-                        to: "lib.rs::shared".to_string(),
-                        is_cycle: false,
-                    },
-                    Edge {
-                        from: "lib.rs::b".to_string(),
-                        to: "lib.rs::shared".to_string(),
-                        is_cycle: false,
-                    },
-                ],
-                roots: vec!["lib.rs::a".to_string(), "lib.rs::b".to_string()],
-            },
-            hotspots: vec![Hotspot {
-                id: "lib.rs::shared".to_string(),
-                path: "lib.rs".to_string(),
-                name: "shared".to_string(),
-                used_by: vec!["a".to_string(), "b".to_string()],
-            }],
-            ..empty_report()
-        };
-
-        let actual = build_detail(&report, "lib.rs::shared").expect("symbol found");
-
-        let mut used_by_names: Vec<&str> = actual.used_by.iter().map(|m| m.name.as_str()).collect();
-        used_by_names.sort();
-        assert_eq!(vec!["a", "b"], used_by_names);
-    }
-
-    #[test]
-    fn should_list_outgoing_edges_as_callees() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![symbol("lib.rs::a", "a"), symbol("lib.rs::b", "b")],
-            }],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("lib.rs::a", "lib.rs", "a"),
-                    node("lib.rs::b", "lib.rs", "b"),
-                ],
-                edges: vec![Edge {
-                    from: "lib.rs::a".to_string(),
-                    to: "lib.rs::b".to_string(),
-                    is_cycle: false,
-                }],
-                roots: vec!["lib.rs::a".to_string()],
-            },
-            ..empty_report()
-        };
-
-        let actual = build_detail(&report, "lib.rs::a").expect("symbol found");
-
-        let expected_callees = vec![SymbolMention {
-            id: "lib.rs::b".to_string(),
-            name: "b".to_string(),
-            path: "lib.rs".to_string(),
-        }];
-        assert_eq!(expected_callees, actual.callees);
-        assert_eq!(Vec::<SymbolMention>::new(), actual.callers);
-    }
-
-    #[test]
-    fn should_have_empty_callers_and_callees_when_symbol_has_no_edges() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![symbol("lib.rs::solo", "solo")],
-            }],
-            graph: SymbolGraph {
-                nodes: vec![node("lib.rs::solo", "lib.rs", "solo")],
-                edges: vec![],
-                roots: vec!["lib.rs::solo".to_string()],
-            },
-            ..empty_report()
-        };
-
-        let actual = build_detail(&report, "lib.rs::solo").expect("symbol found");
-
-        assert_eq!(Vec::<SymbolMention>::new(), actual.callers);
-        assert_eq!(Vec::<SymbolMention>::new(), actual.callees);
-        assert_eq!(Vec::<SymbolMention>::new(), actual.used_by);
-    }
-
-    // SHOULD-FIX 5: `graph.edges` uniqueness is not a contractual guarantee
-    // (`compute_hotspots`'s own doc comment in rinkaku-core::graph notes
-    // this, and defends against it by deduping referrers per target) —
-    // `build_detail` must apply the same defensive dedup to `callees`,
-    // `callers`, and `used_by` rather than assume `graph.edges` never
-    // repeats an edge, or a duplicate edge would silently double-count a
-    // caller/callee in the detail pane's pivots.
-    #[test]
-    fn should_dedup_duplicate_edges_between_the_same_pair_of_symbols() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![
-                    symbol("lib.rs::caller", "caller"),
-                    symbol("lib.rs::callee", "callee"),
-                ],
-            }],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("lib.rs::caller", "lib.rs", "caller"),
-                    node("lib.rs::callee", "lib.rs", "callee"),
-                ],
-                // Same caller -> callee edge listed twice — a hand-built
-                // graph standing in for whatever upstream circumstance
-                // (not currently reachable through `build_graph`, per
-                // `compute_hotspots`'s own doc comment) might one day
-                // produce a repeated edge.
-                edges: vec![
-                    Edge {
-                        from: "lib.rs::caller".to_string(),
-                        to: "lib.rs::callee".to_string(),
-                        is_cycle: false,
-                    },
-                    Edge {
-                        from: "lib.rs::caller".to_string(),
-                        to: "lib.rs::callee".to_string(),
-                        is_cycle: false,
-                    },
-                ],
-                roots: vec!["lib.rs::caller".to_string()],
-            },
-            ..empty_report()
-        };
-
-        let callee_detail = build_detail(&report, "lib.rs::callee").expect("symbol found");
-        let caller_detail = build_detail(&report, "lib.rs::caller").expect("symbol found");
-
-        let expected_caller_mention = vec![SymbolMention {
-            id: "lib.rs::caller".to_string(),
-            name: "caller".to_string(),
-            path: "lib.rs".to_string(),
-        }];
-        let expected_callee_mention = vec![SymbolMention {
-            id: "lib.rs::callee".to_string(),
-            name: "callee".to_string(),
-            path: "lib.rs".to_string(),
-        }];
-        assert_eq!(expected_caller_mention, callee_detail.callers);
-        assert_eq!(expected_caller_mention, callee_detail.used_by);
-        assert_eq!(expected_callee_mention, caller_detail.callees);
-    }
-
-    // SHOULD-FIX 5: `rinkaku-core::graph::collect_edges` explicitly excludes
-    // self-edges (`if target.id != from.id`), so a self-edge cannot occur
-    // through the normal pipeline today — but `build_detail`'s own contract
-    // should not rely on that staying true forever, same reasoning as the
-    // duplicate-edge dedup above. This pins the defensive filter with a
-    // hand-built graph containing a self-edge no real `build_graph` call
-    // would currently produce.
-    #[test]
-    fn should_exclude_self_edge_from_callers_callees_and_used_by() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![symbol("lib.rs::recursive", "recursive")],
-            }],
-            graph: SymbolGraph {
-                nodes: vec![node("lib.rs::recursive", "lib.rs", "recursive")],
-                edges: vec![Edge {
-                    from: "lib.rs::recursive".to_string(),
-                    to: "lib.rs::recursive".to_string(),
-                    is_cycle: false,
-                }],
-                roots: vec!["lib.rs::recursive".to_string()],
-            },
-            ..empty_report()
-        };
-
-        let actual = build_detail(&report, "lib.rs::recursive").expect("symbol found");
-
-        assert_eq!(Vec::<SymbolMention>::new(), actual.callees);
-        assert_eq!(Vec::<SymbolMention>::new(), actual.callers);
-        assert_eq!(Vec::<SymbolMention>::new(), actual.used_by);
-    }
-
-    // symbol_mentions tests (ADR 0022): direct coverage of the function
-    // extracted out of `build_detail`'s own callee/caller computation, so
-    // jump navigation's use of it is pinned independently of the Detail
-    // pane's own tests above.
-
-    #[test]
-    fn should_return_outgoing_edges_as_callees_via_symbol_mentions() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![symbol("lib.rs::a", "a"), symbol("lib.rs::b", "b")],
-            }],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("lib.rs::a", "lib.rs", "a"),
-                    node("lib.rs::b", "lib.rs", "b"),
-                ],
-                edges: vec![Edge {
-                    from: "lib.rs::a".to_string(),
-                    to: "lib.rs::b".to_string(),
-                    is_cycle: false,
-                }],
-                roots: vec!["lib.rs::a".to_string()],
-            },
-            ..empty_report()
-        };
-
-        let actual = symbol_mentions(&report, "lib.rs::a", MentionDirection::Callees);
-
-        assert_eq!(
-            vec![SymbolMention {
-                id: "lib.rs::b".to_string(),
-                name: "b".to_string(),
-                path: "lib.rs".to_string(),
-            }],
-            actual
-        );
-    }
-
-    #[test]
-    fn should_return_incoming_edges_as_callers_via_symbol_mentions() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![symbol("lib.rs::a", "a"), symbol("lib.rs::b", "b")],
-            }],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("lib.rs::a", "lib.rs", "a"),
-                    node("lib.rs::b", "lib.rs", "b"),
-                ],
-                edges: vec![Edge {
-                    from: "lib.rs::a".to_string(),
-                    to: "lib.rs::b".to_string(),
-                    is_cycle: false,
-                }],
-                roots: vec!["lib.rs::a".to_string()],
-            },
-            ..empty_report()
-        };
-
-        let actual = symbol_mentions(&report, "lib.rs::b", MentionDirection::Callers);
-
-        assert_eq!(
-            vec![SymbolMention {
-                id: "lib.rs::a".to_string(),
-                name: "a".to_string(),
-                path: "lib.rs".to_string(),
-            }],
-            actual
-        );
-    }
-
-    #[test]
-    fn should_return_empty_mentions_when_symbol_has_no_edges() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![symbol("lib.rs::solo", "solo")],
-            }],
-            graph: SymbolGraph {
-                nodes: vec![node("lib.rs::solo", "lib.rs", "solo")],
-                edges: vec![],
-                roots: vec!["lib.rs::solo".to_string()],
-            },
-            ..empty_report()
-        };
-
-        let callees = symbol_mentions(&report, "lib.rs::solo", MentionDirection::Callees);
-        let callers = symbol_mentions(&report, "lib.rs::solo", MentionDirection::Callers);
-
-        assert_eq!(Vec::<SymbolMention>::new(), callees);
-        assert_eq!(Vec::<SymbolMention>::new(), callers);
-    }
-
-    // build_dir_detail / build_file_detail tests (TUI iteration 2).
-
-    #[test]
-    fn should_return_none_when_dir_path_is_not_found() {
-        let report = empty_report();
-        let tree = crate::tree::build_tree(&report);
-
-        let actual = build_dir_detail(&tree, &report, "missing");
-
-        assert_eq!(None, actual);
-    }
-
-    #[test]
-    fn should_build_dir_detail_with_badges_and_no_cycle_when_directory_is_not_in_a_cycle() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![ExtractedSymbol {
-                    classification: Some(Classification::SignatureChanged),
-                    ..symbol("src/lib.rs::foo", "foo")
-                }],
-            }],
-            graph: SymbolGraph {
-                nodes: vec![node("src/lib.rs::foo", "src/lib.rs", "foo")],
-                edges: vec![],
-                roots: vec!["src/lib.rs::foo".to_string()],
-            },
-            ..empty_report()
-        };
-        let tree = crate::tree::build_tree(&report);
-
-        let actual = build_dir_detail(&tree, &report, "src").expect("dir found");
-
-        let expected = DirDetail {
-            path: "src".to_string(),
-            badges: crate::tree::Badges {
-                changed_symbols: 1,
-                contract_changes: 1,
-                fan_in: 0,
-                ..crate::tree::Badges::default()
-            },
-            top_fan_in: vec![],
-            cycle_partners: vec![],
-            cycle_edges: vec![],
-        };
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_list_top_fan_in_symbols_sorted_by_fan_in_descending() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![
-                    symbol("src/lib.rs::a", "a"),
-                    symbol("src/lib.rs::b", "b"),
-                    symbol("src/lib.rs::shared_low", "shared_low"),
-                    symbol("src/lib.rs::shared_high", "shared_high"),
-                ],
-            }],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("src/lib.rs::a", "src/lib.rs", "a"),
-                    node("src/lib.rs::b", "src/lib.rs", "b"),
-                    node("src/lib.rs::shared_low", "src/lib.rs", "shared_low"),
-                    node("src/lib.rs::shared_high", "src/lib.rs", "shared_high"),
-                ],
-                edges: vec![],
-                roots: vec![],
-            },
-            hotspots: vec![
-                Hotspot {
-                    id: "src/lib.rs::shared_low".to_string(),
-                    path: "src/lib.rs".to_string(),
-                    name: "shared_low".to_string(),
-                    used_by: vec!["a".to_string(), "b".to_string()],
-                },
-                Hotspot {
-                    id: "src/lib.rs::shared_high".to_string(),
-                    path: "src/lib.rs".to_string(),
-                    name: "shared_high".to_string(),
-                    used_by: vec!["a".to_string(), "b".to_string(), "c".to_string()],
-                },
-            ],
-            ..empty_report()
-        };
-        let tree = crate::tree::build_tree(&report);
-
-        let actual = build_dir_detail(&tree, &report, "src").expect("dir found");
-
-        let expected_top_fan_in = vec![
-            SymbolMention {
-                id: "src/lib.rs::shared_high".to_string(),
-                name: "shared_high".to_string(),
-                path: "src/lib.rs".to_string(),
-            },
-            SymbolMention {
-                id: "src/lib.rs::shared_low".to_string(),
-                name: "shared_low".to_string(),
-                path: "src/lib.rs".to_string(),
-            },
-        ];
-        assert_eq!(expected_top_fan_in, actual.top_fan_in);
-    }
-
-    #[test]
-    fn should_truncate_top_fan_in_to_five_entries() {
-        let symbols: Vec<ExtractedSymbol> = (0..7)
-            .map(|i| symbol(&format!("src/lib.rs::s{i}"), &format!("s{i}")))
-            .collect();
-        let nodes: Vec<Node> = (0..7)
-            .map(|i| node(&format!("src/lib.rs::s{i}"), "src/lib.rs", &format!("s{i}")))
-            .collect();
-        let hotspots: Vec<Hotspot> = (0..7)
-            .map(|i| Hotspot {
-                id: format!("src/lib.rs::s{i}"),
-                path: "src/lib.rs".to_string(),
-                name: format!("s{i}"),
-                used_by: vec!["x".to_string(), "y".to_string()],
-            })
-            .collect();
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols,
-            }],
-            graph: SymbolGraph {
-                nodes,
-                edges: vec![],
-                roots: vec![],
-            },
-            hotspots,
-            ..empty_report()
-        };
-        let tree = crate::tree::build_tree(&report);
-
-        let actual = build_dir_detail(&tree, &report, "src").expect("dir found");
-
-        assert_eq!(5, actual.top_fan_in.len());
-    }
-
-    #[test]
-    fn should_explain_cycle_partners_and_edges_when_directory_participates_in_a_cycle() {
-        // api/ and store/ depend on each other — a directory-level cycle
-        // (mirrors crate::order's own cycle test fixtures).
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![
-                FileReport {
-                    path: "api/handler.rs".to_string(),
-                    symbols: vec![symbol("api/handler.rs::handle", "handle")],
-                },
-                FileReport {
-                    path: "store/db.rs".to_string(),
-                    symbols: vec![symbol("store/db.rs::save", "save")],
-                },
-            ],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("api/handler.rs::handle", "api/handler.rs", "handle"),
-                    node("store/db.rs::save", "store/db.rs", "save"),
-                ],
-                edges: vec![
-                    Edge {
-                        from: "api/handler.rs::handle".to_string(),
-                        to: "store/db.rs::save".to_string(),
-                        is_cycle: false,
-                    },
-                    Edge {
-                        from: "store/db.rs::save".to_string(),
-                        to: "api/handler.rs::handle".to_string(),
-                        is_cycle: false,
-                    },
-                ],
-                roots: vec![],
-            },
-            ..empty_report()
-        };
-        let tree = crate::tree::build_tree(&report);
-
-        let actual = build_dir_detail(&tree, &report, "api").expect("dir found");
-
-        assert_eq!(vec!["store".to_string()], actual.cycle_partners);
-        // Both directed edges touch "api" as an endpoint (api -> store and
-        // store -> api), so both are part of the cycle explanation shown
-        // for "api" — not just the one where "api" is the source.
-        let expected_edges = vec![
-            CycleEdgeView {
-                from: "api/handler.rs::handle".to_string(),
-                to: "store/db.rs::save".to_string(),
-            },
-            CycleEdgeView {
-                from: "store/db.rs::save".to_string(),
-                to: "api/handler.rs::handle".to_string(),
-            },
-        ];
-        assert_eq!(expected_edges, actual.cycle_edges);
-    }
-
-    #[test]
-    fn should_return_none_when_file_path_is_not_found() {
-        let report = empty_report();
-        let tree = crate::tree::build_tree(&report);
-
-        let actual = build_file_detail(&tree, &report, "missing.rs");
-
-        assert_eq!(None, actual);
-    }
-
-    #[test]
-    fn should_build_file_detail_with_symbol_summaries_and_fan_in() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![
-                    ExtractedSymbol {
-                        classification: Some(Classification::Added),
-                        ..symbol("lib.rs::foo", "foo")
-                    },
-                    symbol("lib.rs::bar", "bar"),
-                ],
-            }],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("lib.rs::foo", "lib.rs", "foo"),
-                    node("lib.rs::bar", "lib.rs", "bar"),
-                ],
-                edges: vec![],
-                roots: vec![],
-            },
-            hotspots: vec![Hotspot {
-                id: "lib.rs::bar".to_string(),
-                path: "lib.rs".to_string(),
-                name: "bar".to_string(),
-                used_by: vec!["foo".to_string(), "baz".to_string()],
-            }],
-            ..empty_report()
-        };
-        let tree = crate::tree::build_tree(&report);
-
-        let actual = build_file_detail(&tree, &report, "lib.rs").expect("file found");
-
-        let expected = FileDetail {
-            path: "lib.rs".to_string(),
-            symbols: vec![
-                FileSymbolSummary {
-                    name: "foo".to_string(),
-                    kind: SymbolKind::Function,
-                    classification: Some(Classification::Added),
-                    removed: false,
-                    fan_in: 0,
-                },
-                FileSymbolSummary {
-                    name: "bar".to_string(),
-                    kind: SymbolKind::Function,
-                    classification: None,
-                    removed: false,
-                    fan_in: 2,
-                },
-            ],
-            skip_reason: None,
-            test_symbol_count: None,
-            size_warning: None,
-        };
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_include_removed_symbol_in_file_detail_summary() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![],
-            removed: vec![rinkaku_core::extract::RemovedSymbol {
-                name: "gone".to_string(),
-                kind: SymbolKind::Function,
-                path: "lib.rs".to_string(),
-                signature: "fn gone()".to_string(),
-            }],
-            ..empty_report()
-        };
-        let tree = crate::tree::build_tree(&report);
-
-        let actual = build_file_detail(&tree, &report, "lib.rs").expect("file found");
-
-        let expected = FileDetail {
-            path: "lib.rs".to_string(),
-            symbols: vec![FileSymbolSummary {
-                name: "gone".to_string(),
-                kind: SymbolKind::Function,
-                classification: None,
-                removed: true,
-                fan_in: 0,
-            }],
-            skip_reason: None,
-            test_symbol_count: None,
-            size_warning: None,
-        };
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_carry_skip_reason_into_file_detail_when_file_row_is_skipped() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            skipped: vec![rinkaku_core::render::SkippedFile {
-                path: "assets/logo.png".to_string(),
-                reason: rinkaku_core::render::SkipReason::Binary,
-            }],
-            ..empty_report()
-        };
-        let tree = crate::tree::build_tree(&report);
-
-        let actual = build_file_detail(&tree, &report, "assets/logo.png").expect("file found");
-
-        let expected = FileDetail {
-            path: "assets/logo.png".to_string(),
-            symbols: vec![],
-            skip_reason: Some(rinkaku_core::render::SkipReason::Binary),
-            test_symbol_count: None,
-            size_warning: None,
-        };
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_carry_test_symbol_count_into_file_detail_when_file_row_is_a_whole_test_file() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            tests: vec![rinkaku_core::render::TestFileSummary {
-                path: "src/lib_test.go".to_string(),
-                symbol_count: 4,
-            }],
-            ..empty_report()
-        };
-        let tree = crate::tree::build_tree(&report);
-
-        let actual = build_file_detail(&tree, &report, "src/lib_test.go").expect("file found");
-
-        let expected = FileDetail {
-            path: "src/lib_test.go".to_string(),
-            symbols: vec![],
-            skip_reason: None,
-            test_symbol_count: Some(4),
-            size_warning: None,
-        };
-        assert_eq!(expected, actual);
-    }
-
-    // Regression test (post-rebase integration check): a mixed file — real
-    // (non-test) symbols in `report.files` *and* a test-symbol count in
-    // `report.tests` for the same path, which `pipeline::partition_test_symbols`
-    // legitimately produces for a file with both production and
-    // `#[cfg(test)]`-style code changed in one diff — must keep both halves
-    // on the built `FileDetail` rather than one silently dropping the other.
-    // This is exactly the shape that caused a live panic when running the
-    // TUI against this repo's own diff (`rinkaku-tui/src/app.rs` has both
-    // real and test symbols changed), before `TreeBuilder::insert_test_file`
-    // stopped asserting the file's `symbols` were empty.
-    #[test]
-    fn should_keep_real_symbols_alongside_test_symbol_count_when_file_is_mixed() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "app.rs".to_string(),
-                symbols: vec![symbol("app.rs::handle_key", "handle_key")],
-            }],
-            tests: vec![rinkaku_core::render::TestFileSummary {
-                path: "app.rs".to_string(),
-                symbol_count: 5,
-            }],
-            ..empty_report()
-        };
-        let tree = crate::tree::build_tree(&report);
-
-        let actual = build_file_detail(&tree, &report, "app.rs").expect("file found");
-
-        let expected = FileDetail {
-            path: "app.rs".to_string(),
-            symbols: vec![FileSymbolSummary {
-                name: "handle_key".to_string(),
-                kind: SymbolKind::Function,
-                classification: None,
-                removed: false,
-                fan_in: 0,
-            }],
-            skip_reason: None,
-            test_symbol_count: Some(5),
-            size_warning: None,
-        };
-        assert_eq!(expected, actual);
-    }
-
-    // ADR 0028: a file whose path shows up in `report.file_size_warnings`
-    // must carry the matching warning onto its `FileDetail` so the detail
-    // pane can render the "1734 lines — consider splitting" hint above
-    // the symbols listing without re-scanning the report itself.
-    #[test]
-    fn should_populate_size_warning_on_file_detail_when_report_has_warning_for_that_path() {
-        let warning = rinkaku_core::file_size::FileSizeWarning {
-            path: "src/big.rs".to_string(),
-            line_count: 1734,
-            severity: rinkaku_core::file_size::FileSizeSeverity::Warn,
-        };
-        let report = Report {
-            files: vec![FileReport {
-                path: "src/big.rs".to_string(),
-                symbols: vec![symbol("src/big.rs::foo", "foo")],
-            }],
-            file_size_warnings: vec![warning.clone()],
-            ..empty_report()
-        };
-        let tree = crate::tree::build_tree(&report);
-
-        let actual = build_file_detail(&tree, &report, "src/big.rs").expect("file found");
-
-        let expected = FileDetail {
-            path: "src/big.rs".to_string(),
-            symbols: vec![FileSymbolSummary {
-                name: "foo".to_string(),
-                kind: SymbolKind::Function,
-                classification: None,
-                removed: false,
-                fan_in: 0,
-            }],
-            skip_reason: None,
-            test_symbol_count: None,
-            size_warning: Some(warning),
-        };
-        assert_eq!(expected, actual);
-    }
-}
+#[path = "detail_tests/mod.rs"]
+mod tests;
