+++ rinkaku-tui/src/diff_shape.rs
@@ -317,807 +317,5 @@ fn contract_header_for_symbol(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use pretty_assertions::assert_eq;
-    use rinkaku_core::diff::LineRange;
-    use rinkaku_core::extract::{ExtractedSymbol, SymbolKind};
-    use rinkaku_core::graph::SymbolGraph;
-    use rinkaku_core::render::{FileReport, ReportOrigin};
-
-    fn symbol(id: &str, name: &str, range: LineRange) -> ExtractedSymbol {
-        ExtractedSymbol {
-            id: id.to_string(),
-            name: name.to_string(),
-            kind: SymbolKind::Function,
-            signature: format!("fn {name}()"),
-            range,
-            container: None,
-            referenced_names: vec![],
-            dependencies: vec![],
-            omitted_dependency_matches: 0,
-            is_test: false,
-            classification: None,
-            previous_signature: None,
-        }
-    }
-
-    fn empty_report() -> Report {
-        Report {
-            origin: ReportOrigin::Diff,
-            files: vec![],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![],
-                edges: vec![],
-                roots: vec![],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        }
-    }
-
-    fn hunk(header: &str, new_range: Option<(usize, usize)>, lines: Vec<&str>) -> Hunk {
-        Hunk {
-            header: header.to_string(),
-            new_range,
-            lines: lines
-                .into_iter()
-                .map(|content| crate::diff_view::DiffLine {
-                    kind: crate::diff_view::DiffLineKind::Context,
-                    content: content.to_string(),
-                })
-                .collect(),
-        }
-    }
-
-    /// Wraps `hunk` with the `source_index` it occupies in the fixture's
-    /// `FileHunks::hunks` — every test below builds its `diff_files`
-    /// fixture with hunks in a fixed order, so this index is just "which
-    /// position in that `vec![...]` this hunk was written at".
-    fn attributed(source_index: usize, hunk: Hunk) -> AttributedHunk {
-        AttributedHunk { source_index, hunk }
-    }
-
-    #[test]
-    fn should_return_empty_when_target_is_none() {
-        let report = empty_report();
-
-        let actual = build_diff_pane_content(&report, &[], None);
-
-        assert_eq!(DiffPaneContent::Empty, actual);
-    }
-
-    #[test]
-    fn should_group_file_selection_hunks_under_per_symbol_sections() {
-        let report = Report {
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![
-                    symbol("lib.rs::foo", "foo", LineRange { start: 1, end: 2 }),
-                    symbol("lib.rs::bar", "bar", LineRange { start: 10, end: 11 }),
-                ],
-            }],
-            ..empty_report()
-        };
-        let diff_files = vec![FileHunks {
-            path: "lib.rs".to_string(),
-            hunks: vec![
-                hunk("@@ -1,1 +1,2 @@", Some((1, 2)), vec!["fn foo() {}"]),
-                hunk("@@ -10,1 +10,2 @@", Some((10, 11)), vec!["fn bar() {}"]),
-            ],
-        }];
-        let target = DiffTarget::File {
-            path: "lib.rs".to_string(),
-        };
-
-        let actual = build_diff_pane_content(&report, &diff_files, Some(&target));
-
-        let expected = DiffPaneContent::File(vec![
-            DiffSection {
-                title: "fn foo()".to_string(),
-                symbol_id: Some("lib.rs::foo".to_string()),
-                contract_header: None,
-                hunks: vec![attributed(
-                    0,
-                    hunk("@@ -1,1 +1,2 @@", Some((1, 2)), vec!["fn foo() {}"]),
-                )],
-            },
-            DiffSection {
-                title: "fn bar()".to_string(),
-                symbol_id: Some("lib.rs::bar".to_string()),
-                contract_header: None,
-                hunks: vec![attributed(
-                    1,
-                    hunk("@@ -10,1 +10,2 @@", Some((10, 11)), vec!["fn bar() {}"]),
-                )],
-            },
-        ]);
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_attribute_pure_deletion_hunk_to_owning_symbol_instead_of_module_level() {
-        // Finding-2 regression: `hunk_intersects` always returning `false`
-        // for a pure-deletion hunk meant `build_file_content`'s owner lookup
-        // (`symbols.iter().position(...)`) never matched, so every deletion
-        // hunk landed in the `MODULE_LEVEL_TITLE` bucket regardless of which
-        // symbol's body it actually came from.
-        let report = Report {
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![symbol("lib.rs::foo", "foo", LineRange { start: 1, end: 4 })],
-            }],
-            ..empty_report()
-        };
-        let diff_files = vec![FileHunks {
-            path: "lib.rs".to_string(),
-            hunks: vec![hunk(
-                "@@ -4 +3,0 @@",
-                Some((3, 2)),
-                vec!["println!(\"removed\");"],
-            )],
-        }];
-        let target = DiffTarget::File {
-            path: "lib.rs".to_string(),
-        };
-
-        let actual = build_diff_pane_content(&report, &diff_files, Some(&target));
-
-        let expected = DiffPaneContent::File(vec![DiffSection {
-            title: "fn foo()".to_string(),
-            symbol_id: Some("lib.rs::foo".to_string()),
-            contract_header: None,
-            hunks: vec![attributed(
-                0,
-                hunk(
-                    "@@ -4 +3,0 @@",
-                    Some((3, 2)),
-                    vec!["println!(\"removed\");"],
-                ),
-            )],
-        }]);
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_bucket_hunk_under_module_level_when_it_intersects_no_symbol() {
-        let report = Report {
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![symbol(
-                    "lib.rs::foo",
-                    "foo",
-                    LineRange { start: 10, end: 11 },
-                )],
-            }],
-            ..empty_report()
-        };
-        let diff_files = vec![FileHunks {
-            path: "lib.rs".to_string(),
-            hunks: vec![
-                hunk("@@ -1,1 +1,2 @@", Some((1, 2)), vec!["use foo::bar;"]),
-                hunk("@@ -10,1 +10,2 @@", Some((10, 11)), vec!["fn foo() {}"]),
-            ],
-        }];
-        let target = DiffTarget::File {
-            path: "lib.rs".to_string(),
-        };
-
-        let actual = build_diff_pane_content(&report, &diff_files, Some(&target));
-
-        let expected = DiffPaneContent::File(vec![
-            DiffSection {
-                title: "fn foo()".to_string(),
-                symbol_id: Some("lib.rs::foo".to_string()),
-                contract_header: None,
-                hunks: vec![attributed(
-                    1,
-                    hunk("@@ -10,1 +10,2 @@", Some((10, 11)), vec!["fn foo() {}"]),
-                )],
-            },
-            DiffSection {
-                title: MODULE_LEVEL_TITLE.to_string(),
-                symbol_id: None,
-                contract_header: None,
-                hunks: vec![attributed(
-                    0,
-                    hunk("@@ -1,1 +1,2 @@", Some((1, 2)), vec!["use foo::bar;"]),
-                )],
-            },
-        ]);
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_omit_module_level_section_when_every_hunk_is_attributed_to_a_symbol() {
-        let report = Report {
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![symbol("lib.rs::foo", "foo", LineRange { start: 1, end: 2 })],
-            }],
-            ..empty_report()
-        };
-        let diff_files = vec![FileHunks {
-            path: "lib.rs".to_string(),
-            hunks: vec![hunk("@@ -1,1 +1,2 @@", Some((1, 2)), vec!["fn foo() {}"])],
-        }];
-        let target = DiffTarget::File {
-            path: "lib.rs".to_string(),
-        };
-
-        let actual = build_diff_pane_content(&report, &diff_files, Some(&target));
-
-        let expected = DiffPaneContent::File(vec![DiffSection {
-            title: "fn foo()".to_string(),
-            symbol_id: Some("lib.rs::foo".to_string()),
-            contract_header: None,
-            hunks: vec![attributed(
-                0,
-                hunk("@@ -1,1 +1,2 @@", Some((1, 2)), vec!["fn foo() {}"]),
-            )],
-        }]);
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_attribute_overlapping_hunk_to_every_symbol_it_intersects() {
-        // Two symbols with adjacent, overlapping ranges (a pathological
-        // input a real extractor would not normally produce, but the
-        // shaping function's contract must still resolve deterministically).
-        // ADR 0029 amends ADR 0020's original first-match-only rule: a hunk
-        // intersecting more than one symbol's range is now attributed to
-        // every one of them, not just the first in source order — see ADR
-        // 0029 for why the TUI diff pane departs from ADR 0020's
-        // summary-view "duplication misleads about total change size"
-        // reasoning (the TUI has no change-size total to mislead, and a
-        // dropped section silently breaks that symbol's auto-scroll — ADR
-        // 0027 decision 2 — which is the worse failure mode here).
-        let report = Report {
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![
-                    symbol("lib.rs::foo", "foo", LineRange { start: 1, end: 5 }),
-                    symbol("lib.rs::bar", "bar", LineRange { start: 3, end: 8 }),
-                ],
-            }],
-            ..empty_report()
-        };
-        let diff_files = vec![FileHunks {
-            path: "lib.rs".to_string(),
-            hunks: vec![hunk("@@ -1,1 +1,5 @@", Some((3, 4)), vec!["shared line"])],
-        }];
-        let target = DiffTarget::File {
-            path: "lib.rs".to_string(),
-        };
-
-        let actual = build_diff_pane_content(&report, &diff_files, Some(&target));
-
-        let expected = DiffPaneContent::File(vec![
-            DiffSection {
-                title: "fn foo()".to_string(),
-                symbol_id: Some("lib.rs::foo".to_string()),
-                contract_header: None,
-                hunks: vec![attributed(
-                    0,
-                    hunk("@@ -1,1 +1,5 @@", Some((3, 4)), vec!["shared line"]),
-                )],
-            },
-            DiffSection {
-                title: "fn bar()".to_string(),
-                symbol_id: Some("lib.rs::bar".to_string()),
-                contract_header: None,
-                hunks: vec![attributed(
-                    0,
-                    hunk("@@ -1,1 +1,5 @@", Some((3, 4)), vec!["shared line"]),
-                )],
-            },
-        ]);
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_attribute_new_file_single_hunk_to_every_symbol_it_defines() {
-        // Regression test (PR #86 dogfooding, ADR 0029): a brand-new file's
-        // diff is always exactly one hunk spanning the whole file
-        // (`@@ -0,0 +1,N @@`), so every symbol the file defines has a
-        // range inside that one hunk. Before ADR 0029, only the first
-        // symbol in source order (`foo`) ever got a section; `bar` and
-        // `baz` were silently dropped, breaking their diff-pane auto-scroll
-        // (ADR 0027 decision 2) with no error or indicator.
-        let report = Report {
-            files: vec![FileReport {
-                path: "file_size.rs".to_string(),
-                symbols: vec![
-                    symbol("file_size.rs::foo", "foo", LineRange { start: 1, end: 3 }),
-                    symbol("file_size.rs::bar", "bar", LineRange { start: 5, end: 7 }),
-                    symbol("file_size.rs::baz", "baz", LineRange { start: 9, end: 11 }),
-                ],
-            }],
-            ..empty_report()
-        };
-        let diff_files = vec![FileHunks {
-            path: "file_size.rs".to_string(),
-            hunks: vec![hunk(
-                "@@ -0,0 +1,11 @@",
-                Some((1, 11)),
-                vec!["whole new file"],
-            )],
-        }];
-        let target = DiffTarget::File {
-            path: "file_size.rs".to_string(),
-        };
-
-        let actual = build_diff_pane_content(&report, &diff_files, Some(&target));
-
-        let expected_hunk = attributed(
-            0,
-            hunk("@@ -0,0 +1,11 @@", Some((1, 11)), vec!["whole new file"]),
-        );
-        let expected = DiffPaneContent::File(vec![
-            DiffSection {
-                title: "fn foo()".to_string(),
-                symbol_id: Some("file_size.rs::foo".to_string()),
-                contract_header: None,
-                hunks: vec![expected_hunk.clone()],
-            },
-            DiffSection {
-                title: "fn bar()".to_string(),
-                symbol_id: Some("file_size.rs::bar".to_string()),
-                contract_header: None,
-                hunks: vec![expected_hunk.clone()],
-            },
-            DiffSection {
-                title: "fn baz()".to_string(),
-                symbol_id: Some("file_size.rs::baz".to_string()),
-                contract_header: None,
-                hunks: vec![expected_hunk],
-            },
-        ]);
-        assert_eq!(expected, actual);
-
-        // Every symbol now resolves an auto-scroll target (ADR 0027
-        // decision 2 / decision 4) — not just the first.
-        assert_eq!(
-            Some(0),
-            section_start_line_for_symbol(&actual, "file_size.rs::foo")
-        );
-        assert!(section_start_line_for_symbol(&actual, "file_size.rs::bar").is_some());
-        assert!(section_start_line_for_symbol(&actual, "file_size.rs::baz").is_some());
-    }
-
-    #[test]
-    fn should_include_contract_header_on_the_owning_section_in_a_file_selection() {
-        let report = Report {
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![ExtractedSymbol {
-                    classification: Some(Classification::SignatureChanged),
-                    previous_signature: Some("fn foo(a: i32)".to_string()),
-                    signature: "fn foo(a: i32, b: i32)".to_string(),
-                    ..symbol("lib.rs::foo", "foo", LineRange { start: 1, end: 2 })
-                }],
-            }],
-            ..empty_report()
-        };
-        let diff_files = vec![FileHunks {
-            path: "lib.rs".to_string(),
-            hunks: vec![hunk(
-                "@@ -1,1 +1,2 @@",
-                Some((1, 2)),
-                vec!["fn foo(a, b) {}"],
-            )],
-        }];
-        let target = DiffTarget::File {
-            path: "lib.rs".to_string(),
-        };
-
-        let actual = build_diff_pane_content(&report, &diff_files, Some(&target));
-
-        let expected = DiffPaneContent::File(vec![DiffSection {
-            title: "fn foo(a: i32, b: i32)".to_string(),
-            symbol_id: Some("lib.rs::foo".to_string()),
-            contract_header: Some(ContractHeader {
-                previous_signature: "fn foo(a: i32)".to_string(),
-                signature: "fn foo(a: i32, b: i32)".to_string(),
-            }),
-            hunks: vec![attributed(
-                0,
-                hunk("@@ -1,1 +1,2 @@", Some((1, 2)), vec!["fn foo(a, b) {}"]),
-            )],
-        }]);
-        assert_eq!(expected, actual);
-    }
-
-    // Regression test (post-rebase integration check, PR #58): a skipped or
-    // whole-test-file row (ADR: "show skipped and test-only files in the
-    // entry tree") has no `FileReport` at all in `report.files`, so
-    // `build_file_content`'s `symbols` lookup falls back to `&[]` — every
-    // hunk must still land somewhere rather than being silently dropped or
-    // panicking on an out-of-bounds `sections` index.
-    #[test]
-    fn should_bucket_every_hunk_under_module_level_when_file_selection_has_no_symbols_at_all() {
-        // `report.files` has no entry for "assets/logo.png" at all — the
-        // exact shape of a skipped/whole-test-file row, which is tracked in
-        // `report.skipped`/`report.tests` instead of `report.files`.
-        let report = empty_report();
-        let diff_files = vec![FileHunks {
-            path: "assets/logo.png".to_string(),
-            hunks: vec![hunk(
-                "@@ -1,1 +1,2 @@",
-                Some((1, 2)),
-                vec!["binary blob line"],
-            )],
-        }];
-        let target = DiffTarget::File {
-            path: "assets/logo.png".to_string(),
-        };
-
-        let actual = build_diff_pane_content(&report, &diff_files, Some(&target));
-
-        let expected = DiffPaneContent::File(vec![DiffSection {
-            title: MODULE_LEVEL_TITLE.to_string(),
-            symbol_id: None,
-            contract_header: None,
-            hunks: vec![attributed(
-                0,
-                hunk("@@ -1,1 +1,2 @@", Some((1, 2)), vec!["binary blob line"]),
-            )],
-        }]);
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_return_empty_when_file_has_no_hunks_at_all() {
-        let report = Report {
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![symbol("lib.rs::foo", "foo", LineRange { start: 1, end: 2 })],
-            }],
-            ..empty_report()
-        };
-        let diff_files = vec![FileHunks {
-            path: "lib.rs".to_string(),
-            hunks: vec![],
-        }];
-        let target = DiffTarget::File {
-            path: "lib.rs".to_string(),
-        };
-
-        let actual = build_diff_pane_content(&report, &diff_files, Some(&target));
-
-        assert_eq!(DiffPaneContent::Empty, actual);
-    }
-
-    #[test]
-    fn should_return_empty_when_diff_has_no_entry_for_the_selected_file() {
-        let report = Report {
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![symbol("lib.rs::foo", "foo", LineRange { start: 1, end: 2 })],
-            }],
-            ..empty_report()
-        };
-        let target = DiffTarget::File {
-            path: "lib.rs".to_string(),
-        };
-
-        let actual = build_diff_pane_content(&report, &[], Some(&target));
-
-        assert_eq!(DiffPaneContent::Empty, actual);
-    }
-
-    // Regression test (post-rebase integration check, PR #58): a binary
-    // skipped file has a `FileHunks` entry (git still reports the path
-    // touched a diff) but zero `@@` hunks in it ("Binary files ... differ"
-    // has no hunk syntax for `crate::diff_view::parse_diff_hunks` to parse)
-    // and no `FileReport`/symbols at all — the pane must degrade to `Empty`
-    // (rendered by `crate::ui::draw_diff_pane` as its own placeholder text)
-    // rather than panicking or fabricating a module-level section with no
-    // hunks in it.
-    #[test]
-    fn should_return_empty_when_skipped_file_has_no_symbols_and_no_hunks() {
-        let report = empty_report();
-        let diff_files = vec![FileHunks {
-            path: "assets/logo.png".to_string(),
-            hunks: vec![],
-        }];
-        let target = DiffTarget::File {
-            path: "assets/logo.png".to_string(),
-        };
-
-        let actual = build_diff_pane_content(&report, &diff_files, Some(&target));
-
-        assert_eq!(DiffPaneContent::Empty, actual);
-    }
-
-    // --- hunk_start_lines ---
-
-    #[test]
-    fn should_return_empty_hunk_starts_when_content_is_empty() {
-        let actual = hunk_start_lines(&DiffPaneContent::Empty);
-
-        assert_eq!(Vec::<usize>::new(), actual);
-    }
-
-    #[test]
-    fn should_offset_first_hunk_start_by_title_and_blank_when_file_has_a_single_section_without_contract_header()
-     {
-        // ADR 0027 unified layout: the section title is always shown, and
-        // every hunk (including the section's first) gets a blank line
-        // before its header. So: title(0), blank(1), header(2) — the hunk
-        // starts at line 2.
-        let content = DiffPaneContent::File(vec![DiffSection {
-            title: "fn foo()".to_string(),
-            symbol_id: Some("lib.rs::foo".to_string()),
-            contract_header: None,
-            hunks: vec![attributed(
-                0,
-                hunk(
-                    "@@ -1,1 +1,2 @@",
-                    Some((1, 2)),
-                    vec!["fn a() {}", "fn foo() {}"],
-                ),
-            )],
-        }]);
-
-        let actual = hunk_start_lines(&content);
-
-        assert_eq!(vec![2], actual);
-    }
-
-    #[test]
-    fn should_offset_hunk_start_by_contract_header_lines_when_section_has_one() {
-        // Title(0), 2 contract-header lines (1, 2), blank before the hunk (3),
-        // hunk header (4) — hunk starts at line 4.
-        let content = DiffPaneContent::File(vec![DiffSection {
-            title: "fn foo(a, b)".to_string(),
-            symbol_id: Some("lib.rs::foo".to_string()),
-            contract_header: Some(ContractHeader {
-                previous_signature: "fn foo(a)".to_string(),
-                signature: "fn foo(a, b)".to_string(),
-            }),
-            hunks: vec![attributed(
-                0,
-                hunk("@@ -1,1 +1,2 @@", Some((1, 2)), vec!["fn foo(a, b) {}"]),
-            )],
-        }]);
-
-        let actual = hunk_start_lines(&content);
-
-        assert_eq!(vec![4], actual);
-    }
-
-    #[test]
-    fn should_offset_second_hunk_start_by_first_hunk_header_and_body_length() {
-        // Section title(0), blank(1), first hunk header(2), 2 body lines(3,4),
-        // blank before second hunk(5), second hunk header(6) — starts at 2, 6.
-        let content = DiffPaneContent::File(vec![DiffSection {
-            title: "fn foo()".to_string(),
-            symbol_id: Some("lib.rs::foo".to_string()),
-            contract_header: None,
-            hunks: vec![
-                attributed(
-                    0,
-                    hunk(
-                        "@@ -1,1 +1,2 @@",
-                        Some((1, 2)),
-                        vec!["fn a() {}", "fn b() {}"],
-                    ),
-                ),
-                attributed(
-                    1,
-                    hunk("@@ -10,1 +11,1 @@", Some((11, 11)), vec!["fn c() {}"]),
-                ),
-            ],
-        }]);
-
-        let actual = hunk_start_lines(&content);
-
-        assert_eq!(vec![2, 6], actual);
-    }
-
-    #[test]
-    fn should_offset_hunk_start_by_section_header_and_separator_lines_for_a_file_selection() {
-        // Section 0: title(0), blank(1), header(2), 1 body line(3) — hunk
-        // starts at line 2. Section 1: blank separator between sections(4),
-        // title(5), blank before its hunk(6), header(7) — hunk starts at 7.
-        let content = DiffPaneContent::File(vec![
-            DiffSection {
-                title: "fn foo()".to_string(),
-                symbol_id: Some("lib.rs::foo".to_string()),
-                contract_header: None,
-                hunks: vec![attributed(
-                    0,
-                    hunk("@@ -1,1 +1,2 @@", Some((1, 2)), vec!["fn foo() {}"]),
-                )],
-            },
-            DiffSection {
-                title: "fn bar()".to_string(),
-                symbol_id: Some("lib.rs::bar".to_string()),
-                contract_header: None,
-                hunks: vec![attributed(
-                    1,
-                    hunk("@@ -10,1 +10,2 @@", Some((10, 11)), vec!["fn bar() {}"]),
-                )],
-            },
-        ]);
-
-        let actual = hunk_start_lines(&content);
-
-        assert_eq!(vec![2, 7], actual);
-    }
-
-    #[test]
-    fn should_emit_one_hunk_jump_stop_per_section_when_a_hunk_is_shared_by_two_symbols() {
-        // ADR 0029 consequence: a hunk attributed to more than one symbol
-        // (an overlapping-range case) is rendered once per owning section,
-        // so `]c`/`[c` (backed by this table) must stop once per rendered
-        // occurrence — matching what is actually on screen — rather than
-        // deduplicating by the shared `source_index` down to one stop.
-        // Built straight from `build_diff_pane_content`'s real output
-        // (not a hand-built `DiffPaneContent`) so this test exercises the
-        // same shape the overlapping-hunk attribution test above produces.
-        let report = Report {
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![
-                    symbol("lib.rs::foo", "foo", LineRange { start: 1, end: 5 }),
-                    symbol("lib.rs::bar", "bar", LineRange { start: 3, end: 8 }),
-                ],
-            }],
-            ..empty_report()
-        };
-        let diff_files = vec![FileHunks {
-            path: "lib.rs".to_string(),
-            hunks: vec![hunk("@@ -1,1 +1,5 @@", Some((3, 4)), vec!["shared line"])],
-        }];
-        let target = DiffTarget::File {
-            path: "lib.rs".to_string(),
-        };
-        let content = build_diff_pane_content(&report, &diff_files, Some(&target));
-
-        let actual = hunk_start_lines(&content);
-
-        // Section 0 (`foo`): title(0), blank(1), hunk header(2), 1 body
-        // line(3) — 4 lines. Blank separator(4), section 1 (`bar`)
-        // title(5), blank(6), hunk header(7) — stop at 7. Two stops for
-        // the one underlying hunk, one per section it was duplicated into.
-        assert_eq!(vec![2, 7], actual);
-    }
-
-    // --- section_start_line_for_symbol ---
-
-    #[test]
-    fn should_return_none_for_symbol_start_when_content_is_empty() {
-        let actual = section_start_line_for_symbol(&DiffPaneContent::Empty, "lib.rs::foo");
-
-        assert_eq!(None, actual);
-    }
-
-    #[test]
-    fn should_return_zero_for_symbol_start_when_content_has_a_single_matching_section() {
-        // Only section: its title is at line 0, so the section starts at 0.
-        let content = DiffPaneContent::File(vec![DiffSection {
-            title: "fn foo()".to_string(),
-            symbol_id: Some("lib.rs::foo".to_string()),
-            contract_header: None,
-            hunks: vec![attributed(
-                0,
-                hunk("@@ -1,1 +1,2 @@", Some((1, 2)), vec!["fn foo() {}"]),
-            )],
-        }]);
-
-        let actual = section_start_line_for_symbol(&content, "lib.rs::foo");
-
-        assert_eq!(Some(0), actual);
-    }
-
-    #[test]
-    fn should_return_second_section_start_when_symbol_id_matches_the_second_section() {
-        // Section 0 layout: title(0), blank(1), header(2), body(3) — 4 lines.
-        // Blank separator between sections at line 4, so section 1 starts at 5.
-        let content = DiffPaneContent::File(vec![
-            DiffSection {
-                title: "fn foo()".to_string(),
-                symbol_id: Some("lib.rs::foo".to_string()),
-                contract_header: None,
-                hunks: vec![attributed(
-                    0,
-                    hunk("@@ -1,1 +1,2 @@", Some((1, 2)), vec!["fn foo() {}"]),
-                )],
-            },
-            DiffSection {
-                title: "fn bar()".to_string(),
-                symbol_id: Some("lib.rs::bar".to_string()),
-                contract_header: None,
-                hunks: vec![attributed(
-                    1,
-                    hunk("@@ -10,1 +10,2 @@", Some((10, 11)), vec!["fn bar() {}"]),
-                )],
-            },
-        ]);
-
-        let actual = section_start_line_for_symbol(&content, "lib.rs::bar");
-
-        assert_eq!(Some(5), actual);
-    }
-
-    #[test]
-    fn should_point_at_section_title_line_when_section_has_a_contract_header() {
-        // The section start is the title line, *before* the contract header —
-        // ADR 0027 decision 3: the reviewer wants the section title and its
-        // contract change first, not the hunks below them.
-        // Section 0: title(0), body(1), 2 contract-header lines(2,3), blank
-        // before hunk(4), hunk header(5) — 6 lines. Blank between sections at
-        // line 6, section 1 title at line 7.
-        let content = DiffPaneContent::File(vec![
-            DiffSection {
-                title: "fn foo(a, b)".to_string(),
-                symbol_id: Some("lib.rs::foo".to_string()),
-                contract_header: Some(ContractHeader {
-                    previous_signature: "fn foo(a)".to_string(),
-                    signature: "fn foo(a, b)".to_string(),
-                }),
-                hunks: vec![attributed(
-                    0,
-                    hunk("@@ -1,1 +1,2 @@", Some((1, 2)), vec!["fn foo(a, b) {}"]),
-                )],
-            },
-            DiffSection {
-                title: "fn bar()".to_string(),
-                symbol_id: Some("lib.rs::bar".to_string()),
-                contract_header: None,
-                hunks: vec![attributed(
-                    1,
-                    hunk("@@ -10,1 +10,2 @@", Some((10, 11)), vec!["fn bar() {}"]),
-                )],
-            },
-        ]);
-
-        let actual = section_start_line_for_symbol(&content, "lib.rs::bar");
-
-        assert_eq!(Some(7), actual);
-    }
-
-    #[test]
-    fn should_return_none_for_module_level_bucket_when_asked_by_any_symbol_id() {
-        // The module-level bucket has `symbol_id: None`, so no real symbol id
-        // lookup can accidentally match it. Even passing the literal
-        // `MODULE_LEVEL_TITLE` as a symbol id (which is not a valid symbol id
-        // shape but is the closest a caller could get to "aim at the bucket")
-        // must not match.
-        let content = DiffPaneContent::File(vec![DiffSection {
-            title: MODULE_LEVEL_TITLE.to_string(),
-            symbol_id: None,
-            contract_header: None,
-            hunks: vec![attributed(
-                0,
-                hunk("@@ -1,1 +1,2 @@", Some((1, 2)), vec!["use foo::bar;"]),
-            )],
-        }]);
-
-        let actual = section_start_line_for_symbol(&content, MODULE_LEVEL_TITLE);
-
-        assert_eq!(None, actual);
-    }
-
-    #[test]
-    fn should_return_none_when_symbol_id_matches_no_section() {
-        let content = DiffPaneContent::File(vec![DiffSection {
-            title: "fn foo()".to_string(),
-            symbol_id: Some("lib.rs::foo".to_string()),
-            contract_header: None,
-            hunks: vec![attributed(
-                0,
-                hunk("@@ -1,1 +1,2 @@", Some((1, 2)), vec!["fn foo() {}"]),
-            )],
-        }]);
-
-        let actual = section_start_line_for_symbol(&content, "lib.rs::nonexistent");
-
-        assert_eq!(None, actual);
-    }
-}
+#[path = "diff_shape_tests/mod.rs"]
+mod tests;
