+++ rinkaku-tui/src/order.rs
@@ -633,826 +633,5 @@ fn order_siblings(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use pretty_assertions::assert_eq;
-    use rinkaku_core::extract::{ExtractedSymbol, SymbolKind};
-    use rinkaku_core::graph::{Edge, Node, SymbolGraph};
-    use rinkaku_core::render::FileReport;
-
-    fn symbol(id: &str, name: &str) -> ExtractedSymbol {
-        ExtractedSymbol {
-            id: id.to_string(),
-            name: name.to_string(),
-            kind: SymbolKind::Function,
-            signature: format!("fn {name}()"),
-            range: rinkaku_core::diff::LineRange { start: 1, end: 1 },
-            container: None,
-            referenced_names: vec![],
-            dependencies: vec![],
-            omitted_dependency_matches: 0,
-            is_test: false,
-            classification: None,
-            previous_signature: None,
-        }
-    }
-
-    fn node(id: &str, path: &str, name: &str) -> Node {
-        Node {
-            id: id.to_string(),
-            path: path.to_string(),
-            name: name.to_string(),
-        }
-    }
-
-    fn report_with_graph(nodes: Vec<Node>, edges: Vec<Edge>) -> Report {
-        Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes,
-                edges,
-                roots: vec![],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        }
-    }
-
-    #[test]
-    fn should_return_empty_ranking_when_graph_has_no_nodes() {
-        let report = report_with_graph(vec![], vec![]);
-
-        let expected: HashMap<String, DirRank> = HashMap::new();
-        let actual = rank_directories(&report);
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_rank_directory_with_no_edges_at_rank_zero_and_not_in_cycle() {
-        let report = report_with_graph(vec![node("api/lib.rs::foo", "api/lib.rs", "foo")], vec![]);
-
-        let mut expected = HashMap::new();
-        expected.insert(
-            "api".to_string(),
-            DirRank {
-                rank: 0,
-                in_cycle: false,
-            },
-        );
-        let actual = rank_directories(&report);
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_rank_caller_directory_before_callee_directory() {
-        // api/ calls store/: api is the entry point (0 incoming
-        // inter-directory edges), store is the foundation.
-        let report = report_with_graph(
-            vec![
-                node("api/handler.rs::handle", "api/handler.rs", "handle"),
-                node("store/db.rs::save", "store/db.rs", "save"),
-            ],
-            vec![Edge {
-                from: "api/handler.rs::handle".to_string(),
-                to: "store/db.rs::save".to_string(),
-                is_cycle: false,
-            }],
-        );
-
-        let ranks = rank_directories(&report);
-
-        let api_rank = ranks["api"].rank;
-        let store_rank = ranks["store"].rank;
-        assert!(
-            api_rank < store_rank,
-            "expected api ({api_rank}) to rank before store ({store_rank})"
-        );
-        assert_eq!(false, ranks["api"].in_cycle);
-        assert_eq!(false, ranks["store"].in_cycle);
-    }
-
-    #[test]
-    fn should_drop_edge_and_still_rank_zero_when_both_endpoints_share_a_directory() {
-        // Both symbols live in the same directory ("api"), so the edge
-        // between them condenses to a self-loop and must be dropped —
-        // otherwise "api" would wrongly show up as depending on itself.
-        let report = report_with_graph(
-            vec![
-                node("api/a.rs::a", "api/a.rs", "a"),
-                node("api/b.rs::b", "api/b.rs", "b"),
-            ],
-            vec![Edge {
-                from: "api/a.rs::a".to_string(),
-                to: "api/b.rs::b".to_string(),
-                is_cycle: false,
-            }],
-        );
-
-        let expected = {
-            let mut m = HashMap::new();
-            m.insert(
-                "api".to_string(),
-                DirRank {
-                    rank: 0,
-                    in_cycle: false,
-                },
-            );
-            m
-        };
-        let actual = rank_directories(&report);
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_mark_directories_in_cycle_and_give_them_the_same_rank() {
-        // api/ and store/ depend on each other — a directory-level cycle.
-        let report = report_with_graph(
-            vec![
-                node("api/handler.rs::handle", "api/handler.rs", "handle"),
-                node("store/db.rs::save", "store/db.rs", "save"),
-            ],
-            vec![
-                Edge {
-                    from: "api/handler.rs::handle".to_string(),
-                    to: "store/db.rs::save".to_string(),
-                    is_cycle: false,
-                },
-                Edge {
-                    from: "store/db.rs::save".to_string(),
-                    to: "api/handler.rs::handle".to_string(),
-                    is_cycle: false,
-                },
-            ],
-        );
-
-        let ranks = rank_directories(&report);
-
-        assert_eq!(ranks["api"].rank, ranks["store"].rank);
-        assert_eq!(true, ranks["api"].in_cycle);
-        assert_eq!(true, ranks["store"].in_cycle);
-    }
-
-    #[test]
-    fn should_list_cycle_partners_for_each_directory_in_a_two_directory_cycle() {
-        let report = report_with_graph(
-            vec![
-                node("api/handler.rs::handle", "api/handler.rs", "handle"),
-                node("store/db.rs::save", "store/db.rs", "save"),
-            ],
-            vec![
-                Edge {
-                    from: "api/handler.rs::handle".to_string(),
-                    to: "store/db.rs::save".to_string(),
-                    is_cycle: false,
-                },
-                Edge {
-                    from: "store/db.rs::save".to_string(),
-                    to: "api/handler.rs::handle".to_string(),
-                    is_cycle: false,
-                },
-            ],
-        );
-
-        let partners = cycle_partners(&report);
-
-        let mut expected = HashMap::new();
-        expected.insert("api".to_string(), vec!["store".to_string()]);
-        expected.insert("store".to_string(), vec!["api".to_string()]);
-        assert_eq!(expected, partners);
-    }
-
-    #[test]
-    fn should_return_empty_map_when_no_directory_is_in_a_cycle() {
-        let report = report_with_graph(
-            vec![
-                node("api/a.rs::a", "api/a.rs", "a"),
-                node("store/db.rs::save", "store/db.rs", "save"),
-            ],
-            vec![Edge {
-                from: "api/a.rs::a".to_string(),
-                to: "store/db.rs::save".to_string(),
-                is_cycle: false,
-            }],
-        );
-
-        let partners = cycle_partners(&report);
-
-        let expected: HashMap<String, Vec<String>> = HashMap::new();
-        assert_eq!(expected, partners);
-    }
-
-    #[test]
-    fn should_list_every_partner_when_three_directories_form_one_cycle() {
-        // api -> store -> service -> api: a three-directory cycle, so every
-        // directory's partner list must contain the other two.
-        let report = report_with_graph(
-            vec![
-                node("api/a.rs::a", "api/a.rs", "a"),
-                node("store/s.rs::s", "store/s.rs", "s"),
-                node("service/v.rs::v", "service/v.rs", "v"),
-            ],
-            vec![
-                Edge {
-                    from: "api/a.rs::a".to_string(),
-                    to: "store/s.rs::s".to_string(),
-                    is_cycle: false,
-                },
-                Edge {
-                    from: "store/s.rs::s".to_string(),
-                    to: "service/v.rs::v".to_string(),
-                    is_cycle: false,
-                },
-                Edge {
-                    from: "service/v.rs::v".to_string(),
-                    to: "api/a.rs::a".to_string(),
-                    is_cycle: false,
-                },
-            ],
-        );
-
-        let partners = cycle_partners(&report);
-
-        let mut expected = HashMap::new();
-        expected.insert(
-            "api".to_string(),
-            vec!["service".to_string(), "store".to_string()],
-        );
-        expected.insert(
-            "service".to_string(),
-            vec!["api".to_string(), "store".to_string()],
-        );
-        expected.insert(
-            "store".to_string(),
-            vec!["api".to_string(), "service".to_string()],
-        );
-        assert_eq!(expected, partners);
-    }
-
-    #[test]
-    fn should_return_cross_directory_edges_forming_a_two_directory_cycle() {
-        let report = report_with_graph(
-            vec![
-                node("api/handler.rs::handle", "api/handler.rs", "handle"),
-                node("store/db.rs::save", "store/db.rs", "save"),
-            ],
-            vec![
-                Edge {
-                    from: "api/handler.rs::handle".to_string(),
-                    to: "store/db.rs::save".to_string(),
-                    is_cycle: false,
-                },
-                Edge {
-                    from: "store/db.rs::save".to_string(),
-                    to: "api/handler.rs::handle".to_string(),
-                    is_cycle: false,
-                },
-            ],
-        );
-
-        let mut edges = cycle_edges(&report);
-        edges.sort_by(|a, b| a.from_path.cmp(&b.from_path));
-
-        let expected = vec![
-            CycleEdge {
-                from_path: "api/handler.rs".to_string(),
-                from_name: "handle".to_string(),
-                to_path: "store/db.rs".to_string(),
-                to_name: "save".to_string(),
-            },
-            CycleEdge {
-                from_path: "store/db.rs".to_string(),
-                from_name: "save".to_string(),
-                to_path: "api/handler.rs".to_string(),
-                to_name: "handle".to_string(),
-            },
-        ];
-        assert_eq!(expected, edges);
-    }
-
-    #[test]
-    fn should_exclude_edge_from_cycle_edges_when_directories_do_not_cycle() {
-        // api -> store, but store does not depend back on api: no cycle,
-        // so this edge must not show up as a "cycle edge".
-        let report = report_with_graph(
-            vec![
-                node("api/a.rs::a", "api/a.rs", "a"),
-                node("store/db.rs::save", "store/db.rs", "save"),
-            ],
-            vec![Edge {
-                from: "api/a.rs::a".to_string(),
-                to: "store/db.rs::save".to_string(),
-                is_cycle: false,
-            }],
-        );
-
-        let edges = cycle_edges(&report);
-
-        assert_eq!(Vec::<CycleEdge>::new(), edges);
-    }
-
-    #[test]
-    fn should_exclude_intra_directory_edge_from_cycle_edges() {
-        // Both symbols live in "api" itself; even if api participates in a
-        // cycle with another directory, an edge fully inside api is not a
-        // cross-directory cycle edge.
-        let report = report_with_graph(
-            vec![
-                node("api/a.rs::a", "api/a.rs", "a"),
-                node("api/b.rs::b", "api/b.rs", "b"),
-                node("store/db.rs::save", "store/db.rs", "save"),
-            ],
-            vec![
-                Edge {
-                    from: "api/a.rs::a".to_string(),
-                    to: "api/b.rs::b".to_string(),
-                    is_cycle: false,
-                },
-                Edge {
-                    from: "api/a.rs::a".to_string(),
-                    to: "store/db.rs::save".to_string(),
-                    is_cycle: false,
-                },
-                Edge {
-                    from: "store/db.rs::save".to_string(),
-                    to: "api/a.rs::a".to_string(),
-                    is_cycle: false,
-                },
-            ],
-        );
-
-        let edges = cycle_edges(&report);
-
-        // Only the two cross-directory edges (api <-> store) qualify; the
-        // intra-"api" edge (a -> b) must be excluded.
-        assert_eq!(2, edges.len());
-        assert!(edges.iter().all(|e| e.from_path != e.to_path));
-    }
-
-    #[test]
-    fn should_rank_root_level_file_directory_as_empty_string() {
-        let report = report_with_graph(vec![node("lib.rs::foo", "lib.rs", "foo")], vec![]);
-
-        let expected = {
-            let mut m = HashMap::new();
-            m.insert(
-                String::new(),
-                DirRank {
-                    rank: 0,
-                    in_cycle: false,
-                },
-            );
-            m
-        };
-        let actual = rank_directories(&report);
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_chain_three_directories_in_dependency_order() {
-        // api -> service -> store: three distinct ranks, strictly
-        // increasing in that dependency order.
-        let report = report_with_graph(
-            vec![
-                node("api/a.rs::a", "api/a.rs", "a"),
-                node("service/s.rs::s", "service/s.rs", "s"),
-                node("store/db.rs::save", "store/db.rs", "save"),
-            ],
-            vec![
-                Edge {
-                    from: "api/a.rs::a".to_string(),
-                    to: "service/s.rs::s".to_string(),
-                    is_cycle: false,
-                },
-                Edge {
-                    from: "service/s.rs::s".to_string(),
-                    to: "store/db.rs::save".to_string(),
-                    is_cycle: false,
-                },
-            ],
-        );
-
-        let ranks = rank_directories(&report);
-
-        assert_eq!(0, ranks["api"].rank);
-        assert_eq!(1, ranks["service"].rank);
-        assert_eq!(2, ranks["store"].rank);
-    }
-
-    // Uses `symbol` to document that `rank_directories` only reads
-    // `report.graph`, not `report.files`, even though a realistic `Report`
-    // always carries matching `files` alongside its `graph` — this pins
-    // that `rank_directories`' contract is graph-only, matching its own doc
-    // comment.
-    #[test]
-    fn should_ignore_files_field_and_rank_from_graph_alone() {
-        let mut report = report_with_graph(vec![node("api/a.rs::a", "api/a.rs", "a")], vec![]);
-        report.files = vec![FileReport {
-            path: "api/a.rs".to_string(),
-            symbols: vec![symbol("api/a.rs::a", "a")],
-        }];
-
-        let ranks = rank_directories(&report);
-
-        assert_eq!(1, ranks.len());
-        assert_eq!(0, ranks["api"].rank);
-    }
-
-    fn dir_node(path: &str, children: Vec<crate::tree::TreeNode>) -> crate::tree::TreeNode {
-        crate::tree::TreeNode {
-            kind: crate::tree::NodeKind::Dir,
-            path: path.to_string(),
-            badges: crate::tree::Badges::default(),
-            children,
-            skip_reason: None,
-            test_symbol_count: None,
-        }
-    }
-
-    fn file_node(path: &str) -> crate::tree::TreeNode {
-        crate::tree::TreeNode {
-            kind: crate::tree::NodeKind::File,
-            path: path.to_string(),
-            badges: crate::tree::Badges::default(),
-            children: vec![],
-            skip_reason: None,
-            test_symbol_count: None,
-        }
-    }
-
-    #[test]
-    fn should_order_directories_before_files_at_the_same_level() {
-        let mut tree = crate::tree::Tree {
-            roots: vec![file_node("z.rs"), dir_node("a", vec![])],
-        };
-
-        order_tree(&mut tree, &HashMap::new(), OrderMode::AlphaNumeric);
-
-        let paths: Vec<&str> = tree.roots.iter().map(|n| n.path.as_str()).collect();
-        assert_eq!(vec!["a", "z.rs"], paths);
-    }
-
-    #[test]
-    fn should_order_files_alphabetically_regardless_of_mode() {
-        let mut tree = crate::tree::Tree {
-            roots: vec![file_node("z.rs"), file_node("a.rs")],
-        };
-
-        order_tree(&mut tree, &HashMap::new(), OrderMode::Topological);
-
-        let paths: Vec<&str> = tree.roots.iter().map(|n| n.path.as_str()).collect();
-        assert_eq!(vec!["a.rs", "z.rs"], paths);
-    }
-
-    #[test]
-    fn should_order_directories_by_rank_ascending_when_mode_is_topological() {
-        let mut tree = crate::tree::Tree {
-            roots: vec![dir_node("store", vec![]), dir_node("api", vec![])],
-        };
-        let mut ranks = HashMap::new();
-        ranks.insert(
-            "api".to_string(),
-            DirRank {
-                rank: 0,
-                in_cycle: false,
-            },
-        );
-        ranks.insert(
-            "store".to_string(),
-            DirRank {
-                rank: 1,
-                in_cycle: false,
-            },
-        );
-
-        order_tree(&mut tree, &ranks, OrderMode::Topological);
-
-        let paths: Vec<&str> = tree.roots.iter().map(|n| n.path.as_str()).collect();
-        assert_eq!(vec!["api", "store"], paths);
-    }
-
-    #[test]
-    fn should_order_directories_alphabetically_when_mode_is_alpha_numeric_even_with_ranks_present()
-    {
-        let mut tree = crate::tree::Tree {
-            roots: vec![dir_node("store", vec![]), dir_node("api", vec![])],
-        };
-        let mut ranks = HashMap::new();
-        // Ranks say "store first" but AlphaNumeric mode must ignore them.
-        ranks.insert(
-            "store".to_string(),
-            DirRank {
-                rank: 0,
-                in_cycle: false,
-            },
-        );
-        ranks.insert(
-            "api".to_string(),
-            DirRank {
-                rank: 1,
-                in_cycle: false,
-            },
-        );
-
-        order_tree(&mut tree, &ranks, OrderMode::AlphaNumeric);
-
-        let paths: Vec<&str> = tree.roots.iter().map(|n| n.path.as_str()).collect();
-        assert_eq!(vec!["api", "store"], paths);
-    }
-
-    #[test]
-    fn should_sort_unranked_directory_after_ranked_ones() {
-        let mut tree = crate::tree::Tree {
-            roots: vec![dir_node("unranked", vec![]), dir_node("api", vec![])],
-        };
-        let mut ranks = HashMap::new();
-        ranks.insert(
-            "api".to_string(),
-            DirRank {
-                rank: 0,
-                in_cycle: false,
-            },
-        );
-        // "unranked" has no entry in `ranks` at all.
-
-        order_tree(&mut tree, &ranks, OrderMode::Topological);
-
-        let paths: Vec<&str> = tree.roots.iter().map(|n| n.path.as_str()).collect();
-        assert_eq!(vec!["api", "unranked"], paths);
-    }
-
-    #[test]
-    fn should_order_nested_directory_children_recursively() {
-        let mut tree = crate::tree::Tree {
-            roots: vec![dir_node(
-                "src",
-                vec![dir_node("src/store", vec![]), dir_node("src/api", vec![])],
-            )],
-        };
-        let mut ranks = HashMap::new();
-        ranks.insert(
-            "src/api".to_string(),
-            DirRank {
-                rank: 0,
-                in_cycle: false,
-            },
-        );
-        ranks.insert(
-            "src/store".to_string(),
-            DirRank {
-                rank: 1,
-                in_cycle: false,
-            },
-        );
-
-        order_tree(&mut tree, &ranks, OrderMode::Topological);
-
-        let paths: Vec<&str> = tree.roots[0]
-            .children
-            .iter()
-            .map(|n| n.path.as_str())
-            .collect();
-        assert_eq!(vec!["src/api", "src/store"], paths);
-    }
-
-    // INTEGRATION test: rank_directories -> order_tree end to end, on a
-    // real 3-level branching tree — the shape the unit test above (which
-    // hand-builds the `ranks` map) cannot catch. `rank_directories` only
-    // emits an entry for a *leaf* directory (the direct parent of a graph
-    // node's path); an intermediate/branching directory like "zzz" here
-    // owns no node directly (only its descendants "zzz/api"/"zzz/service"
-    // do). Without effective-rank propagation up through ancestors, "zzz"
-    // would have no rank at all and sort A-Z against "aaa" (i.e. after
-    // it, since "aaa" < "zzz") regardless of dependency direction.
-    //
-    // The whole graph here is a single connected chain (zzz/api ->
-    // zzz/service -> aaa/store) precisely to avoid any ambiguity from
-    // independent/unrelated SCCs sharing in-degree 0 — this test's only
-    // job is to pin down propagation through a branching intermediate
-    // directory, not `topological_scc_order`'s tie-break among unrelated
-    // components (covered separately, see the `tarjan_sccs`/
-    // `topological_scc_order` unit tests below). Package names are chosen
-    // so the correct topological order ("zzz" first, "aaa" last) disagrees
-    // with alphabetical order, making a regression to A-Z observable.
-    #[test]
-    fn should_order_full_tree_by_effective_rank_through_branching_intermediate_directories() {
-        // zzz/api/handler.rs -> zzz/service/logic.rs -> aaa/store/db.rs:
-        // "zzz" contains both the entry point (api, rank 0) and a middle
-        // link (service, rank 1) — its effective rank must become 0 (the
-        // minimum of its descendants) so it still sorts first despite
-        // "aaa" < "zzz" alphabetically.
-        let report = report_with_graph(
-            vec![
-                node("zzz/api/handler.rs::handle", "zzz/api/handler.rs", "handle"),
-                node("zzz/service/logic.rs::run", "zzz/service/logic.rs", "run"),
-                node("aaa/store/db.rs::save", "aaa/store/db.rs", "save"),
-            ],
-            vec![
-                Edge {
-                    from: "zzz/api/handler.rs::handle".to_string(),
-                    to: "zzz/service/logic.rs::run".to_string(),
-                    is_cycle: false,
-                },
-                Edge {
-                    from: "zzz/service/logic.rs::run".to_string(),
-                    to: "aaa/store/db.rs::save".to_string(),
-                    is_cycle: false,
-                },
-            ],
-        );
-        let ranks = rank_directories(&report);
-
-        // Build the corresponding tree by hand (mirrors what
-        // `tree::build_tree` would produce for this Report's files, minus
-        // badges which this test doesn't care about): two top-level
-        // packages, "zzz" branching into two subdirectories, "aaa" with
-        // one.
-        let mut tree = crate::tree::Tree {
-            roots: vec![
-                dir_node(
-                    "aaa",
-                    vec![dir_node(
-                        "aaa/store",
-                        vec![file_node_with_children("aaa/store/db.rs", vec![])],
-                    )],
-                ),
-                dir_node(
-                    "zzz",
-                    vec![
-                        dir_node(
-                            "zzz/service",
-                            vec![file_node_with_children("zzz/service/logic.rs", vec![])],
-                        ),
-                        dir_node(
-                            "zzz/api",
-                            vec![file_node_with_children("zzz/api/handler.rs", vec![])],
-                        ),
-                    ],
-                ),
-            ],
-        };
-
-        order_tree(&mut tree, &ranks, OrderMode::Topological);
-
-        let expected = crate::tree::Tree {
-            roots: vec![
-                dir_node(
-                    "zzz",
-                    vec![
-                        dir_node(
-                            "zzz/api",
-                            vec![file_node_with_children("zzz/api/handler.rs", vec![])],
-                        ),
-                        dir_node(
-                            "zzz/service",
-                            vec![file_node_with_children("zzz/service/logic.rs", vec![])],
-                        ),
-                    ],
-                ),
-                dir_node(
-                    "aaa",
-                    vec![dir_node(
-                        "aaa/store",
-                        vec![file_node_with_children("aaa/store/db.rs", vec![])],
-                    )],
-                ),
-            ],
-        };
-        assert_eq!(expected, tree);
-    }
-
-    fn file_node_with_children(
-        path: &str,
-        children: Vec<crate::tree::TreeNode>,
-    ) -> crate::tree::TreeNode {
-        crate::tree::TreeNode {
-            kind: crate::tree::NodeKind::File,
-            path: path.to_string(),
-            badges: crate::tree::Badges::default(),
-            children,
-            skip_reason: None,
-            test_symbol_count: None,
-        }
-    }
-
-    // The `rank_directories`-level tests above exercise `tarjan_sccs` and
-    // `topological_scc_order` only indirectly, through a full directory
-    // condensation built from a `Report`. These tests instead drive the two
-    // private helpers directly with hand-built adjacency lists, so a future
-    // change to either algorithm gets pinned down at the shape it actually
-    // operates on (plain `usize` adjacency), not just observed through the
-    // directory-name-shaped output several layers up.
-
-    #[test]
-    fn should_split_into_singleton_sccs_when_graph_is_a_diamond() {
-        // 0 -> 1, 0 -> 2, 1 -> 3, 2 -> 3: acyclic diamond, so every node is
-        // its own SCC (no back edges to merge any of them together).
-        let adjacency = vec![vec![1, 2], vec![3], vec![3], vec![]];
-
-        let mut sccs = tarjan_sccs(&adjacency);
-        for scc in &mut sccs {
-            scc.sort_unstable();
-        }
-        sccs.sort_by_key(|scc| scc[0]);
-
-        assert_eq!(vec![vec![0], vec![1], vec![2], vec![3]], sccs);
-    }
-
-    #[test]
-    fn should_group_mutually_reachable_nodes_into_one_scc_when_graph_has_two_independent_cycles() {
-        // 0 <-> 1 and 2 <-> 3, with no edges between the two pairs: two
-        // independent 2-node SCCs, neither depending on the other.
-        let adjacency = vec![vec![1], vec![0], vec![3], vec![2]];
-
-        let mut sccs = tarjan_sccs(&adjacency);
-        for scc in &mut sccs {
-            scc.sort_unstable();
-        }
-        sccs.sort_by_key(|scc| scc[0]);
-
-        assert_eq!(vec![vec![0, 1], vec![2, 3]], sccs);
-    }
-
-    #[test]
-    fn should_isolate_cycle_from_surrounding_acyclic_nodes_when_graph_mixes_both() {
-        // 0 -> 1 -> 2 -> 1 -> 3: nodes 1 and 2 form a cycle (1 -> 2 -> 1),
-        // while 0 (entry) and 3 (reached only from the cycle) stay acyclic
-        // singletons on either side of it.
-        let adjacency = vec![vec![1], vec![2], vec![1, 3], vec![]];
-
-        let mut sccs = tarjan_sccs(&adjacency);
-        for scc in &mut sccs {
-            scc.sort_unstable();
-        }
-        sccs.sort_by_key(|scc| scc[0]);
-
-        assert_eq!(vec![vec![0], vec![1, 2], vec![3]], sccs);
-    }
-
-    #[test]
-    fn should_order_diamond_sccs_with_entry_first_and_join_last() {
-        // Same diamond as the tarjan_sccs test above: 0 -> {1, 2} -> 3.
-        // Every node is its own singleton SCC; topological order must place
-        // 0 first (the only in-degree-0 SCC) and 3 last (depended on by
-        // both 1 and 2), with 1 before 2 as the deterministic tie-break
-        // (smaller member index first) since neither depends on the other.
-        let adjacency = vec![vec![1, 2], vec![3], vec![3], vec![]];
-        let sccs: Vec<Vec<usize>> = vec![vec![0], vec![1], vec![2], vec![3]];
-        let mut scc_of = vec![0usize; 4];
-        for (scc_index, scc) in sccs.iter().enumerate() {
-            for &node_index in scc {
-                scc_of[node_index] = scc_index;
-            }
-        }
-
-        let order = topological_scc_order(&sccs, &scc_of, &adjacency);
-
-        assert_eq!(vec![0, 1, 2, 3], order);
-    }
-
-    #[test]
-    fn should_order_two_independent_sccs_by_smallest_member_when_neither_depends_on_the_other() {
-        // {0, 1} and {2, 3} are independent SCCs (no edges between them) —
-        // both have in-degree 0 in the condensation, so the tie-break
-        // (smallest member index) alone decides that {0, 1} orders first.
-        let adjacency = vec![vec![1], vec![0], vec![3], vec![2]];
-        let sccs: Vec<Vec<usize>> = vec![vec![0, 1], vec![2, 3]];
-        let mut scc_of = vec![0usize; 4];
-        for (scc_index, scc) in sccs.iter().enumerate() {
-            for &node_index in scc {
-                scc_of[node_index] = scc_index;
-            }
-        }
-
-        let order = topological_scc_order(&sccs, &scc_of, &adjacency);
-
-        assert_eq!(vec![0, 1], order);
-    }
-
-    #[test]
-    fn should_order_cycle_between_its_acyclic_neighbors_when_graph_mixes_both() {
-        // Same mixed graph as the tarjan_sccs test above: 0 -> {1, 2} (a
-        // cycle) -> 3. The condensation is 0 -> scc{1,2} -> 3, so the
-        // topological order must be exactly [0, scc{1,2}, 3] regardless of
-        // which of 1/2 tarjan_sccs happened to list first inside the SCC.
-        let adjacency = vec![vec![1], vec![2], vec![1, 3], vec![]];
-        let sccs: Vec<Vec<usize>> = vec![vec![0], vec![1, 2], vec![3]];
-        let mut scc_of = vec![0usize; 4];
-        for (scc_index, scc) in sccs.iter().enumerate() {
-            for &node_index in scc {
-                scc_of[node_index] = scc_index;
-            }
-        }
-
-        let order = topological_scc_order(&sccs, &scc_of, &adjacency);
-
-        assert_eq!(vec![0, 1, 2], order);
-    }
-}
+#[path = "order_tests/mod.rs"]
+mod tests;
