+++ rinkaku-tui/src/row_view.rs
@@ -401,697 +401,5 @@ fn kind_abbrev(kind: SymbolKind) -> &'static str {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::tree::{NodeKind, TreeNode};
-
-    fn dir_node(path: &str, badges: Badges, children: Vec<TreeNode>) -> TreeNode {
-        TreeNode {
-            kind: NodeKind::Dir,
-            path: path.to_string(),
-            badges,
-            children,
-            skip_reason: None,
-            test_symbol_count: None,
-        }
-    }
-
-    fn file_node(path: &str, badges: Badges) -> TreeNode {
-        TreeNode {
-            kind: NodeKind::File,
-            path: path.to_string(),
-            badges,
-            children: vec![],
-            skip_reason: None,
-            test_symbol_count: None,
-        }
-    }
-
-    fn skipped_file_node(path: &str, reason: rinkaku_core::render::SkipReason) -> TreeNode {
-        TreeNode {
-            kind: NodeKind::File,
-            path: path.to_string(),
-            badges: Badges::default(),
-            children: vec![],
-            skip_reason: Some(reason),
-            test_symbol_count: None,
-        }
-    }
-
-    fn test_file_node(path: &str, symbol_count: usize) -> TreeNode {
-        TreeNode {
-            kind: NodeKind::File,
-            path: path.to_string(),
-            badges: Badges::default(),
-            children: vec![],
-            skip_reason: None,
-            test_symbol_count: Some(symbol_count),
-        }
-    }
-
-    fn symbol_node(path: &str, symbol_ref: SymbolRef, badges: Badges) -> TreeNode {
-        TreeNode {
-            kind: NodeKind::Symbol(symbol_ref),
-            path: path.to_string(),
-            badges,
-            children: vec![],
-            skip_reason: None,
-            test_symbol_count: None,
-        }
-    }
-
-    fn plain_symbol(name: &str) -> SymbolRef {
-        SymbolRef {
-            id: format!("lib.rs::{name}"),
-            name: name.to_string(),
-            kind: SymbolKind::Function,
-            classification: None,
-            removed: false,
-        }
-    }
-
-    fn line_text(line: &Line<'_>) -> String {
-        line.spans
-            .iter()
-            .map(|span| span.content.as_ref())
-            .collect()
-    }
-
-    #[test]
-    fn should_render_plain_text_for_zero_badges_and_no_classification() {
-        let node = symbol_node("lib.rs", plain_symbol("foo"), Badges::default());
-        let row = Row {
-            node: &node,
-            depth: 2,
-            expanded: false,
-        };
-
-        let line = entry_row_line(&row, "", &HashMap::new(), false);
-
-        assert_eq!("        fn foo", line_text(&line));
-    }
-
-    #[test]
-    fn should_include_badge_labels_for_nonzero_badges_on_a_dir_row() {
-        // ADR 0013 amendment (2026-07-13): the changed-symbol and fan-in
-        // badges use `chg:` / `ref:` text labels instead of the original
-        // `~` / `^` glyphs. `!{N}` (contract-change count) is
-        // intentionally left as a compact glyph — see `push_badge_spans`'
-        // doc comment for the scope split.
-        let node = dir_node(
-            "src",
-            Badges {
-                changed_symbols: 2,
-                contract_changes: 1,
-                fan_in: 3,
-                ..Badges::default()
-            },
-            vec![file_node("src/a.rs", Badges::default())],
-        );
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: true,
-        };
-
-        let line = entry_row_line(&row, "src", &HashMap::new(), false);
-
-        assert_eq!("v src chg:2 !1 ref:3", line_text(&line));
-    }
-
-    #[test]
-    fn should_omit_zero_badges_entirely() {
-        let node = file_node("lib.rs", Badges::default());
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: false,
-        };
-
-        let line = entry_row_line(&row, "lib.rs", &HashMap::new(), false);
-
-        assert_eq!("  lib.rs ", line_text(&line));
-    }
-
-    #[test]
-    fn should_append_skip_reason_for_a_skipped_file_row() {
-        let node = skipped_file_node("assets/logo.png", rinkaku_core::render::SkipReason::Binary);
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: false,
-        };
-
-        let line = entry_row_line(&row, "assets/logo.png", &HashMap::new(), false);
-
-        assert_eq!("  assets/logo.png  (skipped: binary)", line_text(&line));
-    }
-
-    #[test]
-    fn should_dim_label_for_a_skipped_file_row() {
-        let node = skipped_file_node("assets/logo.png", rinkaku_core::render::SkipReason::Binary);
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: false,
-        };
-
-        let line = entry_row_line(&row, "assets/logo.png", &HashMap::new(), false);
-
-        // The label span is the third span: indent, expand marker, label.
-        assert_eq!(Some(Color::DarkGray), line.spans[2].style.fg);
-    }
-
-    #[test]
-    fn should_not_append_skip_reason_for_an_ordinary_file_row() {
-        let node = file_node("lib.rs", Badges::default());
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: false,
-        };
-
-        let line = entry_row_line(&row, "lib.rs", &HashMap::new(), false);
-
-        assert!(!line_text(&line).contains("skipped"));
-    }
-
-    #[test]
-    fn should_append_test_badge_with_plural_symbols_noun_for_a_whole_test_file_row() {
-        let node = test_file_node("src/lib_test.go", 3);
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: false,
-        };
-
-        let line = entry_row_line(&row, "src/lib_test.go", &HashMap::new(), false);
-
-        assert_eq!("  src/lib_test.go  [test] (3 symbols)", line_text(&line));
-    }
-
-    #[test]
-    fn should_append_test_badge_with_singular_symbol_noun_when_count_is_one() {
-        let node = test_file_node("src/lib_test.go", 1);
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: false,
-        };
-
-        let line = entry_row_line(&row, "src/lib_test.go", &HashMap::new(), false);
-
-        assert_eq!("  src/lib_test.go  [test] (1 symbol)", line_text(&line));
-    }
-
-    #[test]
-    fn should_show_collapse_marker_when_dir_is_not_expanded() {
-        let node = dir_node(
-            "src",
-            Badges::default(),
-            vec![file_node("src/a.rs", Badges::default())],
-        );
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: false,
-        };
-
-        let line = entry_row_line(&row, "src", &HashMap::new(), false);
-
-        assert_eq!("> src ", line_text(&line));
-    }
-
-    #[test]
-    fn should_append_cycle_marker_when_dir_path_is_in_cycle() {
-        let node = dir_node("src", Badges::default(), vec![]);
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: false,
-        };
-        let mut ranks = HashMap::new();
-        ranks.insert(
-            "src".to_string(),
-            DirRank {
-                rank: 0,
-                in_cycle: true,
-            },
-        );
-
-        let line = entry_row_line(&row, "src", &ranks, false);
-
-        assert_eq!("  src  (cycle)", line_text(&line));
-    }
-
-    #[test]
-    fn should_not_append_cycle_marker_when_dir_path_is_not_in_cycle() {
-        let node = dir_node("src", Badges::default(), vec![]);
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: false,
-        };
-        let mut ranks = HashMap::new();
-        ranks.insert(
-            "src".to_string(),
-            DirRank {
-                rank: 0,
-                in_cycle: false,
-            },
-        );
-
-        let line = entry_row_line(&row, "src", &ranks, false);
-
-        assert_eq!("  src ", line_text(&line));
-    }
-
-    #[test]
-    fn should_mark_added_symbol_with_plus() {
-        let symbol_ref = SymbolRef {
-            classification: Some(Classification::Added),
-            ..plain_symbol("new_fn")
-        };
-        let node = symbol_node("lib.rs", symbol_ref, Badges::default());
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: false,
-        };
-
-        let line = entry_row_line(&row, "", &HashMap::new(), false);
-
-        assert_eq!("  + fn new_fn", line_text(&line));
-    }
-
-    #[test]
-    fn should_mark_signature_changed_symbol_with_tilde() {
-        let symbol_ref = SymbolRef {
-            classification: Some(Classification::SignatureChanged),
-            ..plain_symbol("changed_fn")
-        };
-        let node = symbol_node("lib.rs", symbol_ref, Badges::default());
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: false,
-        };
-
-        let line = entry_row_line(&row, "", &HashMap::new(), false);
-
-        assert_eq!("  ~ fn changed_fn", line_text(&line));
-    }
-
-    #[test]
-    fn should_mark_removed_symbol_with_x() {
-        let symbol_ref = SymbolRef {
-            removed: true,
-            ..plain_symbol("gone_fn")
-        };
-        let node = symbol_node("lib.rs", symbol_ref, Badges::default());
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: false,
-        };
-
-        let line = entry_row_line(&row, "", &HashMap::new(), false);
-
-        assert_eq!("  x fn gone_fn", line_text(&line));
-    }
-
-    #[test]
-    fn should_apply_reversed_modifier_when_row_is_selected() {
-        let node = file_node("lib.rs", Badges::default());
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: false,
-        };
-
-        let line = entry_row_line(&row, "lib.rs", &HashMap::new(), true);
-
-        assert!(line.style.add_modifier.contains(Modifier::REVERSED));
-    }
-
-    #[test]
-    fn should_not_apply_reversed_modifier_when_row_is_not_selected() {
-        let node = file_node("lib.rs", Badges::default());
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: false,
-        };
-
-        let line = entry_row_line(&row, "lib.rs", &HashMap::new(), false);
-
-        assert!(!line.style.add_modifier.contains(Modifier::REVERSED));
-    }
-
-    #[test]
-    fn should_indent_by_depth_times_indent_width() {
-        let node = file_node("lib.rs", Badges::default());
-        let row = Row {
-            node: &node,
-            depth: 3,
-            expanded: false,
-        };
-
-        let line = entry_row_line(&row, "lib.rs", &HashMap::new(), false);
-
-        assert_eq!("        lib.rs ", line_text(&line));
-    }
-
-    #[test]
-    fn should_strip_ancestor_prefix_from_nested_dir_label() {
-        // "src" (depth 0) then "src/foo" (depth 1): the second row's label
-        // must be just "foo", not the full "src/foo" its `path` carries.
-        let root = dir_node("src", Badges::default(), vec![]);
-        let child = dir_node("src/foo", Badges::default(), vec![]);
-        let rows = vec![
-            Row {
-                node: &root,
-                depth: 0,
-                expanded: true,
-            },
-            Row {
-                node: &child,
-                depth: 1,
-                expanded: false,
-            },
-        ];
-
-        let labels = relative_labels(&rows);
-
-        assert_eq!(vec!["src".to_string(), "foo".to_string()], labels);
-    }
-
-    #[test]
-    fn should_keep_full_collapsed_label_when_no_ancestor_row_precedes_it() {
-        // A root-level collapsed chain "src/foo/bar" has no ancestor row
-        // above it at all, so its label stays the full collapsed path.
-        let root = dir_node("src/foo/bar", Badges::default(), vec![]);
-        let rows = vec![Row {
-            node: &root,
-            depth: 0,
-            expanded: true,
-        }];
-
-        let labels = relative_labels(&rows);
-
-        assert_eq!(vec!["src/foo/bar".to_string()], labels);
-    }
-
-    #[test]
-    fn should_not_strip_partial_string_overlap_between_sibling_directory_names() {
-        // "src" and "src2" are two independent top-level roots (both
-        // depth 0, i.e. siblings, not ancestor/descendant) that happen to
-        // share "src" as a string prefix. `relative_labels` only ever
-        // compares a row against its own ancestor chain (via
-        // `ancestor_path_at[row.depth - 1]`), never against a sibling at
-        // the same depth, so "src2" must keep its full label rather than
-        // having "src" spuriously stripped off it as if "src" were its
-        // parent.
-        let src = dir_node("src", Badges::default(), vec![]);
-        let src2 = dir_node("src2", Badges::default(), vec![]);
-        let rows = vec![
-            Row {
-                node: &src,
-                depth: 0,
-                expanded: false,
-            },
-            Row {
-                node: &src2,
-                depth: 0,
-                expanded: false,
-            },
-        ];
-
-        let labels = relative_labels(&rows);
-
-        assert_eq!(vec!["src".to_string(), "src2".to_string()], labels);
-    }
-
-    #[test]
-    fn should_return_empty_label_for_symbol_rows() {
-        let node = symbol_node("lib.rs", plain_symbol("foo"), Badges::default());
-        let rows = vec![Row {
-            node: &node,
-            depth: 1,
-            expanded: false,
-        }];
-
-        let labels = relative_labels(&rows);
-
-        assert_eq!(vec![String::new()], labels);
-    }
-
-    #[test]
-    fn should_recompute_label_correctly_after_returning_to_a_shallower_sibling() {
-        // "a" (depth 0) -> "a/one.rs" (depth 1) -> "b" (depth 0, a sibling
-        // of "a", not a descendant) -> "b/two.rs" (depth 1). The second
-        // "depth 1" row must strip "b"'s prefix, not stale ancestor state
-        // left over from "a"'s subtree.
-        let a = dir_node("a", Badges::default(), vec![]);
-        let a_file = file_node("a/one.rs", Badges::default());
-        let b = dir_node("b", Badges::default(), vec![]);
-        let b_file = file_node("b/two.rs", Badges::default());
-        let rows = vec![
-            Row {
-                node: &a,
-                depth: 0,
-                expanded: true,
-            },
-            Row {
-                node: &a_file,
-                depth: 1,
-                expanded: false,
-            },
-            Row {
-                node: &b,
-                depth: 0,
-                expanded: true,
-            },
-            Row {
-                node: &b_file,
-                depth: 1,
-                expanded: false,
-            },
-        ];
-
-        let labels = relative_labels(&rows);
-
-        assert_eq!(
-            vec![
-                "a".to_string(),
-                "one.rs".to_string(),
-                "b".to_string(),
-                "two.rs".to_string(),
-            ],
-            labels
-        );
-    }
-
-    // File-size warning badges (ADR 0028): a file row carrying a warning
-    // renders `lines:{N}` with the numeric N colored by severity (yellow
-    // for Warn, red for Split), and a directory row aggregates the
-    // per-severity counts as `warn:N split:N` with the numbers colored
-    // the same way. No emoji glyphs — the color already conveys severity
-    // and emoji rendering width is inconsistent across terminals.
-
-    /// Locates the styled span whose visible text is `content` in `line`,
-    /// returning its foreground color (or `None` when the span exists but
-    /// has no explicit fg, matching ratatui's default). Panics when no
-    /// such span exists — a matching-span assertion is what the test
-    /// wanted, so a missing span is a test failure, not a `None`.
-    fn fg_of_span_with_content(line: &Line<'_>, content: &str) -> Option<Color> {
-        line.spans
-            .iter()
-            .find(|span| span.content.as_ref() == content)
-            .unwrap_or_else(|| {
-                panic!(
-                    "no span with content {content:?} found in line: {:?}",
-                    line_text(line)
-                )
-            })
-            .style
-            .fg
-    }
-
-    #[test]
-    fn should_render_lines_count_in_yellow_when_file_has_warn_severity() {
-        let node = TreeNode {
-            kind: NodeKind::File,
-            path: "src/big.rs".to_string(),
-            badges: Badges {
-                own_file_size_severity: Some(FileSizeSeverity::Warn),
-                own_file_line_count: Some(1734),
-                ..Badges::default()
-            },
-            children: vec![],
-            skip_reason: None,
-            test_symbol_count: None,
-        };
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: false,
-        };
-
-        let line = entry_row_line(&row, "big.rs", &HashMap::new(), false);
-
-        assert_eq!("  big.rs lines:1734", line_text(&line));
-        // The numeric 1734 span carries the severity color; the leading
-        // "lines:" label stays uncolored so the eye lands on the number.
-        assert_eq!(Some(Color::Yellow), fg_of_span_with_content(&line, "1734"));
-        assert_eq!(None, fg_of_span_with_content(&line, "lines:"));
-    }
-
-    #[test]
-    fn should_render_lines_count_in_red_when_file_has_split_severity() {
-        let node = TreeNode {
-            kind: NodeKind::File,
-            path: "src/huge.rs".to_string(),
-            badges: Badges {
-                own_file_size_severity: Some(FileSizeSeverity::Split),
-                own_file_line_count: Some(4837),
-                ..Badges::default()
-            },
-            children: vec![],
-            skip_reason: None,
-            test_symbol_count: None,
-        };
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: false,
-        };
-
-        let line = entry_row_line(&row, "huge.rs", &HashMap::new(), false);
-
-        assert_eq!("  huge.rs lines:4837", line_text(&line));
-        assert_eq!(Some(Color::Red), fg_of_span_with_content(&line, "4837"));
-    }
-
-    #[test]
-    fn should_render_warn_and_split_labels_side_by_side_on_dir_row() {
-        let node = dir_node(
-            "src",
-            Badges {
-                file_size_warn_count: 2,
-                file_size_split_count: 1,
-                ..Badges::default()
-            },
-            vec![file_node("src/a.rs", Badges::default())],
-        );
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: true,
-        };
-
-        let line = entry_row_line(&row, "src", &HashMap::new(), false);
-
-        assert_eq!("v src warn:2 split:1", line_text(&line));
-        // The numeric part of each half picks up its own severity color;
-        // the "warn:" / "split:" labels themselves stay uncolored so the
-        // eye lands on the counts.
-        assert_eq!(Some(Color::Yellow), fg_of_span_with_content(&line, "2"));
-        assert_eq!(Some(Color::Red), fg_of_span_with_content(&line, "1"));
-        assert_eq!(None, fg_of_span_with_content(&line, "warn:"));
-        assert_eq!(None, fg_of_span_with_content(&line, "split:"));
-    }
-
-    #[test]
-    fn should_not_render_file_size_badge_when_file_node_has_no_warning() {
-        let node = file_node("lib.rs", Badges::default());
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: false,
-        };
-
-        let line = entry_row_line(&row, "lib.rs", &HashMap::new(), false);
-
-        let text = line_text(&line);
-        assert!(!text.contains("lines:"));
-        assert!(!text.contains("warn:"));
-        assert!(!text.contains("split:"));
-    }
-
-    #[test]
-    fn should_render_only_warn_label_on_dir_row_when_no_split_files() {
-        // When only one severity is present under a directory, only that
-        // half of the badge shows — the other half is omitted rather
-        // than rendered as "warn:0" or "split:0".
-        let node = dir_node(
-            "src",
-            Badges {
-                file_size_warn_count: 3,
-                file_size_split_count: 0,
-                ..Badges::default()
-            },
-            vec![file_node("src/a.rs", Badges::default())],
-        );
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: true,
-        };
-
-        let line = entry_row_line(&row, "src", &HashMap::new(), false);
-
-        assert_eq!("v src warn:3", line_text(&line));
-    }
-
-    // ADR 0013 amendment (2026-07-13): `chg:N` and `ref:N` badges split
-    // their label from their number across two spans so only the number
-    // picks up cyan — matching the file-size badges' split-span pattern
-    // (`lines:N`, `warn:N`, `split:N`). The label prefix reads at the
-    // default color to keep the eye on the numeric part.
-    #[test]
-    fn should_color_only_the_number_of_chg_badge_and_leave_label_uncolored() {
-        let node = dir_node(
-            "src",
-            Badges {
-                changed_symbols: 299,
-                ..Badges::default()
-            },
-            vec![file_node("src/a.rs", Badges::default())],
-        );
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: true,
-        };
-
-        let line = entry_row_line(&row, "src", &HashMap::new(), false);
-
-        assert_eq!("v src chg:299", line_text(&line));
-        assert_eq!(Some(Color::Cyan), fg_of_span_with_content(&line, "299"));
-        assert_eq!(None, fg_of_span_with_content(&line, "chg:"));
-    }
-
-    #[test]
-    fn should_color_only_the_number_of_ref_badge_and_leave_label_uncolored() {
-        let node = dir_node(
-            "src",
-            Badges {
-                fan_in: 1072,
-                ..Badges::default()
-            },
-            vec![file_node("src/a.rs", Badges::default())],
-        );
-        let row = Row {
-            node: &node,
-            depth: 0,
-            expanded: true,
-        };
-
-        let line = entry_row_line(&row, "src", &HashMap::new(), false);
-
-        assert_eq!("v src ref:1072", line_text(&line));
-        assert_eq!(Some(Color::Cyan), fg_of_span_with_content(&line, "1072"));
-        assert_eq!(None, fg_of_span_with_content(&line, "ref:"));
-    }
-}
+#[path = "row_view_tests/mod.rs"]
+mod tests;
