+++ rinkaku-tui/src/row_view_tests/chg_ref_badges.rs
@@ -1,0 +1,53 @@
+// ADR 0013 amendment (2026-07-13): `chg:N` and `ref:N` badges split
+// their label from their number across two spans so only the number
+// picks up cyan — matching the file-size badges' split-span pattern
+// (`lines:N`, `warn:N`, `split:N`). The label prefix reads at the
+// default color to keep the eye on the numeric part.
+
+use super::*;
+
+#[test]
+fn should_color_only_the_number_of_chg_badge_and_leave_label_uncolored() {
+    let node = dir_node(
+        "src",
+        Badges {
+            changed_symbols: 299,
+            ..Badges::default()
+        },
+        vec![file_node("src/a.rs", Badges::default())],
+    );
+    let row = Row {
+        node: &node,
+        depth: 0,
+        expanded: true,
+    };
+
+    let line = entry_row_line(&row, "src", &HashMap::new(), false);
+
+    assert_eq!("v src chg:299", line_text(&line));
+    assert_eq!(Some(Color::Cyan), fg_of_span_with_content(&line, "299"));
+    assert_eq!(None, fg_of_span_with_content(&line, "chg:"));
+}
+
+#[test]
+fn should_color_only_the_number_of_ref_badge_and_leave_label_uncolored() {
+    let node = dir_node(
+        "src",
+        Badges {
+            fan_in: 1072,
+            ..Badges::default()
+        },
+        vec![file_node("src/a.rs", Badges::default())],
+    );
+    let row = Row {
+        node: &node,
+        depth: 0,
+        expanded: true,
+    };
+
+    let line = entry_row_line(&row, "src", &HashMap::new(), false);
+
+    assert_eq!("v src ref:1072", line_text(&line));
+    assert_eq!(Some(Color::Cyan), fg_of_span_with_content(&line, "1072"));
+    assert_eq!(None, fg_of_span_with_content(&line, "ref:"));
+}
