+++ rinkaku-tui/src/tree.rs
@@ -642,979 +642,5 @@ fn join_path(prefix: &str, segment: &str) -> String {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use pretty_assertions::assert_eq;
-    use rinkaku_core::diff::LineRange;
-    use rinkaku_core::extract::{ExtractedSymbol, RemovedSymbol};
-    use rinkaku_core::graph::{Hotspot, SymbolGraph};
-    use rinkaku_core::render::{FileReport, SkippedFile, TestFileSummary};
-
-    fn symbol(id: &str, name: &str, kind: SymbolKind) -> ExtractedSymbol {
-        ExtractedSymbol {
-            id: id.to_string(),
-            name: name.to_string(),
-            kind,
-            signature: format!("fn {name}()"),
-            range: LineRange { start: 1, end: 1 },
-            container: None,
-            referenced_names: vec![],
-            dependencies: vec![],
-            omitted_dependency_matches: 0,
-            is_test: false,
-            classification: None,
-            previous_signature: None,
-        }
-    }
-
-    fn empty_report() -> Report {
-        Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![],
-                edges: vec![],
-                roots: vec![],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        }
-    }
-
-    #[test]
-    fn should_build_empty_tree_when_report_has_no_files_and_no_removed() {
-        let report = empty_report();
-
-        let expected = Tree { roots: vec![] };
-        let actual = build_tree(&report);
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_build_flat_file_node_when_path_has_no_directory() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![symbol("lib.rs::foo", "foo", SymbolKind::Function)],
-            }],
-            ..empty_report()
-        };
-
-        let expected = Tree {
-            roots: vec![TreeNode {
-                kind: NodeKind::File,
-                path: "lib.rs".to_string(),
-                badges: Badges {
-                    changed_symbols: 1,
-                    contract_changes: 0,
-                    fan_in: 0,
-                    ..Badges::default()
-                },
-                children: vec![TreeNode {
-                    kind: NodeKind::Symbol(SymbolRef {
-                        id: "lib.rs::foo".to_string(),
-                        name: "foo".to_string(),
-                        kind: SymbolKind::Function,
-                        classification: None,
-                        removed: false,
-                    }),
-                    path: "lib.rs".to_string(),
-                    badges: Badges {
-                        changed_symbols: 1,
-                        contract_changes: 0,
-                        fan_in: 0,
-                        ..Badges::default()
-                    },
-                    children: vec![],
-                    skip_reason: None,
-                    test_symbol_count: None,
-                }],
-                skip_reason: None,
-                test_symbol_count: None,
-            }],
-        };
-        let actual = build_tree(&report);
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_collapse_single_child_directory_chain_into_one_node() {
-        // src/foo/bar/lib.rs — src, foo, bar each have exactly one child,
-        // so all three collapse into one Dir node labeled "src/foo/bar".
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/foo/bar/lib.rs".to_string(),
-                symbols: vec![],
-            }],
-            ..empty_report()
-        };
-
-        let expected = Tree {
-            roots: vec![TreeNode {
-                kind: NodeKind::Dir,
-                path: "src/foo/bar".to_string(),
-                badges: Badges::default(),
-                children: vec![TreeNode {
-                    kind: NodeKind::File,
-                    path: "src/foo/bar/lib.rs".to_string(),
-                    badges: Badges::default(),
-                    children: vec![],
-                    skip_reason: None,
-                    test_symbol_count: None,
-                }],
-                skip_reason: None,
-                test_symbol_count: None,
-            }],
-        };
-        let actual = build_tree(&report);
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_not_collapse_directory_with_two_children() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![
-                FileReport {
-                    path: "src/a.rs".to_string(),
-                    symbols: vec![],
-                },
-                FileReport {
-                    path: "src/b.rs".to_string(),
-                    symbols: vec![],
-                },
-            ],
-            ..empty_report()
-        };
-
-        let expected = Tree {
-            roots: vec![TreeNode {
-                kind: NodeKind::Dir,
-                path: "src".to_string(),
-                badges: Badges::default(),
-                children: vec![
-                    TreeNode {
-                        kind: NodeKind::File,
-                        path: "src/a.rs".to_string(),
-                        badges: Badges::default(),
-                        children: vec![],
-                        skip_reason: None,
-                        test_symbol_count: None,
-                    },
-                    TreeNode {
-                        kind: NodeKind::File,
-                        path: "src/b.rs".to_string(),
-                        badges: Badges::default(),
-                        children: vec![],
-                        skip_reason: None,
-                        test_symbol_count: None,
-                    },
-                ],
-                skip_reason: None,
-                test_symbol_count: None,
-            }],
-        };
-        let actual = build_tree(&report);
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_not_collapse_directory_that_has_own_file_alongside_subdirectory() {
-        // src/ has both a direct file (mod.rs) and a subdirectory (foo/) —
-        // src is not "just a chain" to reach foo, so it must stay a
-        // separate node rather than collapsing with foo.
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![
-                FileReport {
-                    path: "src/mod.rs".to_string(),
-                    symbols: vec![],
-                },
-                FileReport {
-                    path: "src/foo/bar.rs".to_string(),
-                    symbols: vec![],
-                },
-            ],
-            ..empty_report()
-        };
-
-        let expected = Tree {
-            roots: vec![TreeNode {
-                kind: NodeKind::Dir,
-                path: "src".to_string(),
-                badges: Badges::default(),
-                children: vec![
-                    TreeNode {
-                        kind: NodeKind::File,
-                        path: "src/mod.rs".to_string(),
-                        badges: Badges::default(),
-                        children: vec![],
-                        skip_reason: None,
-                        test_symbol_count: None,
-                    },
-                    TreeNode {
-                        kind: NodeKind::Dir,
-                        path: "src/foo".to_string(),
-                        badges: Badges::default(),
-                        children: vec![TreeNode {
-                            kind: NodeKind::File,
-                            path: "src/foo/bar.rs".to_string(),
-                            badges: Badges::default(),
-                            children: vec![],
-                            skip_reason: None,
-                            test_symbol_count: None,
-                        }],
-                        skip_reason: None,
-                        test_symbol_count: None,
-                    },
-                ],
-                skip_reason: None,
-                test_symbol_count: None,
-            }],
-        };
-        let actual = build_tree(&report);
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_count_contract_change_for_signature_changed_symbol() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![ExtractedSymbol {
-                    classification: Some(Classification::SignatureChanged),
-                    ..symbol("lib.rs::foo", "foo", SymbolKind::Function)
-                }],
-            }],
-            ..empty_report()
-        };
-
-        let tree = build_tree(&report);
-
-        let expected = Badges {
-            changed_symbols: 1,
-            contract_changes: 1,
-            fan_in: 0,
-            ..Badges::default()
-        };
-        let actual = tree.roots[0].badges;
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_not_count_contract_change_for_body_only_symbol() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![ExtractedSymbol {
-                    classification: Some(Classification::BodyOnly),
-                    ..symbol("lib.rs::foo", "foo", SymbolKind::Function)
-                }],
-            }],
-            ..empty_report()
-        };
-
-        let tree = build_tree(&report);
-
-        let expected = Badges {
-            changed_symbols: 1,
-            contract_changes: 0,
-            fan_in: 0,
-            ..Badges::default()
-        };
-        let actual = tree.roots[0].badges;
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_add_removed_symbol_as_marked_leaf_under_its_file_without_counting_as_changed() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![],
-            removed: vec![RemovedSymbol {
-                name: "gone".to_string(),
-                kind: SymbolKind::Function,
-                path: "lib.rs".to_string(),
-                signature: "fn gone()".to_string(),
-            }],
-            ..empty_report()
-        };
-
-        let expected = Tree {
-            roots: vec![TreeNode {
-                kind: NodeKind::File,
-                path: "lib.rs".to_string(),
-                badges: Badges {
-                    changed_symbols: 0,
-                    contract_changes: 1,
-                    fan_in: 0,
-                    ..Badges::default()
-                },
-                children: vec![TreeNode {
-                    kind: NodeKind::Symbol(SymbolRef {
-                        id: "lib.rs::gone".to_string(),
-                        name: "gone".to_string(),
-                        kind: SymbolKind::Function,
-                        classification: None,
-                        removed: true,
-                    }),
-                    path: "lib.rs".to_string(),
-                    badges: Badges {
-                        changed_symbols: 0,
-                        contract_changes: 1,
-                        fan_in: 0,
-                        ..Badges::default()
-                    },
-                    children: vec![],
-                    skip_reason: None,
-                    test_symbol_count: None,
-                }],
-                skip_reason: None,
-                test_symbol_count: None,
-            }],
-        };
-        let actual = build_tree(&report);
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_merge_removed_symbol_into_existing_file_with_present_symbols() {
-        // A file with one present (unchanged classification-wise) symbol
-        // and one removed symbol must land under the same File node, not
-        // create two separate entries for "lib.rs".
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![symbol("lib.rs::foo", "foo", SymbolKind::Function)],
-            }],
-            removed: vec![RemovedSymbol {
-                name: "gone".to_string(),
-                kind: SymbolKind::Function,
-                path: "lib.rs".to_string(),
-                signature: "fn gone()".to_string(),
-            }],
-            ..empty_report()
-        };
-
-        let expected = Tree {
-            roots: vec![TreeNode {
-                kind: NodeKind::File,
-                path: "lib.rs".to_string(),
-                badges: Badges {
-                    changed_symbols: 1,
-                    contract_changes: 1,
-                    fan_in: 0,
-                    ..Badges::default()
-                },
-                children: vec![
-                    TreeNode {
-                        kind: NodeKind::Symbol(SymbolRef {
-                            id: "lib.rs::foo".to_string(),
-                            name: "foo".to_string(),
-                            kind: SymbolKind::Function,
-                            classification: None,
-                            removed: false,
-                        }),
-                        path: "lib.rs".to_string(),
-                        badges: Badges {
-                            changed_symbols: 1,
-                            contract_changes: 0,
-                            fan_in: 0,
-                            ..Badges::default()
-                        },
-                        children: vec![],
-                        skip_reason: None,
-                        test_symbol_count: None,
-                    },
-                    TreeNode {
-                        kind: NodeKind::Symbol(SymbolRef {
-                            id: "lib.rs::gone".to_string(),
-                            name: "gone".to_string(),
-                            kind: SymbolKind::Function,
-                            classification: None,
-                            removed: true,
-                        }),
-                        path: "lib.rs".to_string(),
-                        badges: Badges {
-                            changed_symbols: 0,
-                            contract_changes: 1,
-                            fan_in: 0,
-                            ..Badges::default()
-                        },
-                        children: vec![],
-                        skip_reason: None,
-                        test_symbol_count: None,
-                    },
-                ],
-                skip_reason: None,
-                test_symbol_count: None,
-            }],
-        };
-        let actual = build_tree(&report);
-
-        assert_eq!(expected, actual);
-    }
-
-    // NOTE: partial assert (root count, path, then only the aggregated
-    // `Badges`) rather than a whole-`Tree` comparison — this test's only
-    // concern is that bottom-up aggregation reaches the top of a
-    // multi-level, multi-file subtree correctly; restating the full
-    // "src/a/one.rs" and "src/b/two.rs" node structure (already pinned down
-    // by other tests in this module, e.g.
-    // `should_build_flat_file_node_when_path_has_no_directory`) would just
-    // add noise without strengthening what this test is checking.
-    #[test]
-    fn should_aggregate_badges_bottom_up_across_nested_directories() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![
-                FileReport {
-                    path: "src/a/one.rs".to_string(),
-                    symbols: vec![ExtractedSymbol {
-                        classification: Some(Classification::SignatureChanged),
-                        ..symbol("src/a/one.rs::x", "x", SymbolKind::Function)
-                    }],
-                },
-                FileReport {
-                    path: "src/b/two.rs".to_string(),
-                    symbols: vec![symbol("src/b/two.rs::y", "y", SymbolKind::Function)],
-                },
-            ],
-            ..empty_report()
-        };
-
-        let tree = build_tree(&report);
-
-        assert_eq!(1, tree.roots.len());
-        let src = &tree.roots[0];
-        assert_eq!("src", src.path);
-        let expected = Badges {
-            changed_symbols: 2,
-            contract_changes: 1,
-            fan_in: 0,
-            ..Badges::default()
-        };
-        assert_eq!(expected, src.badges);
-    }
-
-    #[test]
-    fn should_keep_file_with_no_symbols_as_childless_file_node() {
-        // A pure rename (FileReport with empty symbols) must still show up
-        // as a File node with zero badges, not be dropped from the tree.
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/renamed.rs".to_string(),
-                symbols: vec![],
-            }],
-            ..empty_report()
-        };
-
-        let expected = Tree {
-            roots: vec![TreeNode {
-                kind: NodeKind::Dir,
-                path: "src".to_string(),
-                badges: Badges::default(),
-                children: vec![TreeNode {
-                    kind: NodeKind::File,
-                    path: "src/renamed.rs".to_string(),
-                    badges: Badges::default(),
-                    children: vec![],
-                    skip_reason: None,
-                    test_symbol_count: None,
-                }],
-                skip_reason: None,
-                test_symbol_count: None,
-            }],
-        };
-        let actual = build_tree(&report);
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_preserve_source_order_of_siblings_before_reordering() {
-        // Discovery order in `report.files` must be preserved (reordering
-        // is a separate concern handled by `crate::order`), even though the
-        // builder uses a BTreeMap internally.
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![
-                FileReport {
-                    path: "z.rs".to_string(),
-                    symbols: vec![],
-                },
-                FileReport {
-                    path: "a.rs".to_string(),
-                    symbols: vec![],
-                },
-            ],
-            ..empty_report()
-        };
-
-        let tree = build_tree(&report);
-
-        let names: Vec<&str> = tree.roots.iter().map(|n| n.path.as_str()).collect();
-        assert_eq!(vec!["z.rs", "a.rs"], names);
-    }
-
-    #[test]
-    fn should_set_fan_in_badge_from_matching_hotspot_and_aggregate_upward() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![symbol("src/lib.rs::shared", "shared", SymbolKind::Function)],
-            }],
-            hotspots: vec![Hotspot {
-                id: "src/lib.rs::shared".to_string(),
-                path: "src/lib.rs".to_string(),
-                name: "shared".to_string(),
-                used_by: vec!["a".to_string(), "b".to_string()],
-            }],
-            ..empty_report()
-        };
-
-        let tree = build_tree(&report);
-
-        // Fan-in of 2 (two referrers) must show on the symbol leaf and
-        // aggregate up through File and Dir.
-        let src = &tree.roots[0];
-        assert_eq!("src", src.path);
-        assert_eq!(2, src.badges.fan_in);
-        let file_node = &src.children[0];
-        assert_eq!(2, file_node.badges.fan_in);
-        let symbol_node = &file_node.children[0];
-        assert_eq!(2, symbol_node.badges.fan_in);
-    }
-
-    #[test]
-    fn should_leave_fan_in_at_zero_when_symbol_has_no_matching_hotspot() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![symbol("lib.rs::solo", "solo", SymbolKind::Function)],
-            }],
-            hotspots: vec![],
-            ..empty_report()
-        };
-
-        let tree = build_tree(&report);
-
-        assert_eq!(0, tree.roots[0].badges.fan_in);
-    }
-
-    // File-size warning badge tests (ADR 0028): a file rinkaku measured as
-    // oversized must carry its own severity + line count on the file node,
-    // and its containing directories must aggregate the count split by
-    // severity so `row_view` can render the pair (`warn:N split:M`).
-
-    #[test]
-    fn should_populate_own_file_line_count_when_report_has_warning_for_that_path() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/big.rs".to_string(),
-                symbols: vec![],
-            }],
-            file_size_warnings: vec![rinkaku_core::file_size::FileSizeWarning {
-                path: "src/big.rs".to_string(),
-                line_count: 1734,
-                severity: FileSizeSeverity::Warn,
-            }],
-            ..empty_report()
-        };
-
-        let tree = build_tree(&report);
-
-        // Root is the "src" Dir; file node is its only child.
-        let file_node = &tree.roots[0].children[0];
-        let expected = Badges {
-            own_file_size_severity: Some(FileSizeSeverity::Warn),
-            own_file_line_count: Some(1734),
-            file_size_warn_count: 1,
-            file_size_split_count: 0,
-            ..Badges::default()
-        };
-        assert_eq!(expected, file_node.badges);
-    }
-
-    #[test]
-    fn should_aggregate_warn_and_split_counts_separately_on_parent_dir() {
-        // src/warn.rs = Warn, src/split.rs = Split, src/ok.rs = under
-        // threshold. The parent "src" dir must show warn_count=1,
-        // split_count=1 — kept split by severity so the row can render
-        // `warn:1 split:1` rather than merging into one meaningless total.
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![
-                FileReport {
-                    path: "src/warn.rs".to_string(),
-                    symbols: vec![],
-                },
-                FileReport {
-                    path: "src/split.rs".to_string(),
-                    symbols: vec![],
-                },
-                FileReport {
-                    path: "src/ok.rs".to_string(),
-                    symbols: vec![],
-                },
-            ],
-            file_size_warnings: vec![
-                rinkaku_core::file_size::FileSizeWarning {
-                    path: "src/warn.rs".to_string(),
-                    line_count: 1600,
-                    severity: FileSizeSeverity::Warn,
-                },
-                rinkaku_core::file_size::FileSizeWarning {
-                    path: "src/split.rs".to_string(),
-                    line_count: 2500,
-                    severity: FileSizeSeverity::Split,
-                },
-            ],
-            ..empty_report()
-        };
-
-        let tree = build_tree(&report);
-
-        // Root is the "src" Dir.
-        let expected = Badges {
-            file_size_warn_count: 1,
-            file_size_split_count: 1,
-            // Directories deliberately never carry per-node severity /
-            // line_count of their own — those live only on file rows.
-            own_file_size_severity: None,
-            own_file_line_count: None,
-            ..Badges::default()
-        };
-        assert_eq!(expected, tree.roots[0].badges);
-    }
-
-    #[test]
-    fn should_leave_own_file_size_severity_none_on_dir_node() {
-        // Even when a directory aggregates a nonzero warn/split count
-        // from below, its own_file_size_severity / own_file_line_count
-        // must stay None — those two fields are per-file attributes only.
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/big.rs".to_string(),
-                symbols: vec![],
-            }],
-            file_size_warnings: vec![rinkaku_core::file_size::FileSizeWarning {
-                path: "src/big.rs".to_string(),
-                line_count: 3000,
-                severity: FileSizeSeverity::Split,
-            }],
-            ..empty_report()
-        };
-
-        let tree = build_tree(&report);
-
-        let dir_node = &tree.roots[0];
-        assert_eq!(None, dir_node.badges.own_file_size_severity);
-        assert_eq!(None, dir_node.badges.own_file_line_count);
-        // Aggregates still count the descendant.
-        assert_eq!(1, dir_node.badges.file_size_split_count);
-    }
-
-    // Skipped-file tests: a file rinkaku could not extract symbols from
-    // (unsupported language, binary, deleted) must still show up in the
-    // tree, since otherwise it is invisible to a reviewer relying on the
-    // TUI to see the whole PR (the user-reported gap this feature closes).
-
-    #[test]
-    fn should_add_skipped_file_as_childless_file_node_with_skip_reason() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            skipped: vec![SkippedFile {
-                path: "assets/logo.png".to_string(),
-                reason: rinkaku_core::render::SkipReason::Binary,
-            }],
-            ..empty_report()
-        };
-
-        let expected = Tree {
-            roots: vec![TreeNode {
-                kind: NodeKind::Dir,
-                path: "assets".to_string(),
-                badges: Badges::default(),
-                children: vec![TreeNode {
-                    kind: NodeKind::File,
-                    path: "assets/logo.png".to_string(),
-                    badges: Badges::default(),
-                    children: vec![],
-                    skip_reason: Some(rinkaku_core::render::SkipReason::Binary),
-                    test_symbol_count: None,
-                }],
-                skip_reason: None,
-                test_symbol_count: None,
-            }],
-        };
-        let actual = build_tree(&report);
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_omit_generated_skip_reason_from_tree_by_default() {
-        // Mirrors `render_markdown`'s own `SkipReason::Generated` filter
-        // (ADR 0010/0011): a `.gitattributes`-declared or content-marked
-        // generated file is already known-uninteresting, so it should not
-        // clutter the TUI tree either.
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            skipped: vec![SkippedFile {
-                path: "Cargo.lock".to_string(),
-                reason: rinkaku_core::render::SkipReason::Generated,
-            }],
-            ..empty_report()
-        };
-
-        let tree = build_tree(&report);
-
-        assert_eq!(Tree { roots: vec![] }, tree);
-    }
-
-    #[test]
-    fn should_keep_non_generated_skip_reasons_when_mixed_with_a_generated_entry() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            skipped: vec![
-                SkippedFile {
-                    path: "Cargo.lock".to_string(),
-                    reason: rinkaku_core::render::SkipReason::Generated,
-                },
-                SkippedFile {
-                    path: "assets/logo.png".to_string(),
-                    reason: rinkaku_core::render::SkipReason::Binary,
-                },
-            ],
-            ..empty_report()
-        };
-
-        let tree = build_tree(&report);
-
-        let paths: Vec<&str> = tree.roots.iter().map(|n| n.path.as_str()).collect();
-        assert_eq!(vec!["assets"], paths);
-    }
-
-    #[test]
-    fn should_merge_skipped_file_into_existing_dir_alongside_analyzed_files() {
-        // A skipped file sharing a directory with an already-analyzed file
-        // must land in the same `Dir` node, not create a second "src" root.
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![],
-            }],
-            skipped: vec![SkippedFile {
-                path: "src/generated.pb.go".to_string(),
-                reason: rinkaku_core::render::SkipReason::UnsupportedLanguage,
-            }],
-            ..empty_report()
-        };
-
-        let tree = build_tree(&report);
-
-        assert_eq!(1, tree.roots.len());
-        let src = &tree.roots[0];
-        assert_eq!("src", src.path);
-        assert_eq!(2, src.children.len());
-        // Source order: `report.files` is inserted before `report.skipped`
-        // (see `build_tree`'s own doc comment), so the analyzed file comes
-        // first even though "generated.pb.go" sorts first alphabetically.
-        let paths: Vec<&str> = src.children.iter().map(|n| n.path.as_str()).collect();
-        assert_eq!(vec!["src/lib.rs", "src/generated.pb.go"], paths);
-    }
-
-    // Whole-test-file tests: a file whose changed symbols were *all* test
-    // code has no `FileReport` in `report.files` at all
-    // (`pipeline::partition_test_symbols`'s doc comment) — only a
-    // `TestFileSummary` in `report.tests`. Without surfacing that summary
-    // into the tree, such a file is invisible to a reviewer, the same gap
-    // as a skipped file.
-
-    #[test]
-    fn should_add_whole_test_file_as_childless_file_node_with_symbol_count() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            tests: vec![TestFileSummary {
-                path: "src/lib_test.go".to_string(),
-                symbol_count: 3,
-            }],
-            ..empty_report()
-        };
-
-        let expected = Tree {
-            roots: vec![TreeNode {
-                kind: NodeKind::Dir,
-                path: "src".to_string(),
-                badges: Badges::default(),
-                children: vec![TreeNode {
-                    kind: NodeKind::File,
-                    path: "src/lib_test.go".to_string(),
-                    badges: Badges::default(),
-                    children: vec![],
-                    skip_reason: None,
-                    test_symbol_count: Some(3),
-                }],
-                skip_reason: None,
-                test_symbol_count: None,
-            }],
-        };
-        let actual = build_tree(&report);
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_merge_test_file_into_existing_dir_alongside_analyzed_files() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![symbol("src/lib.rs::foo", "foo", SymbolKind::Function)],
-            }],
-            tests: vec![TestFileSummary {
-                path: "src/lib_test.rs".to_string(),
-                symbol_count: 2,
-            }],
-            ..empty_report()
-        };
-
-        let tree = build_tree(&report);
-
-        assert_eq!(1, tree.roots.len());
-        let src = &tree.roots[0];
-        assert_eq!("src", src.path);
-        assert_eq!(2, src.children.len());
-        let paths: Vec<&str> = src.children.iter().map(|n| n.path.as_str()).collect();
-        assert_eq!(vec!["src/lib.rs", "src/lib_test.rs"], paths);
-    }
-
-    #[test]
-    fn should_not_set_test_symbol_count_or_skip_reason_on_an_ordinary_file() {
-        // Regression guard: an ordinary analyzed file must keep both new
-        // fields at `None`, not accidentally inherit a stale default from
-        // whatever `FileBuilder` construction path is taken.
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![],
-            }],
-            ..empty_report()
-        };
-
-        let tree = build_tree(&report);
-
-        assert_eq!(None, tree.roots[0].skip_reason);
-        assert_eq!(None, tree.roots[0].test_symbol_count);
-    }
-
-    // `pipeline::analyze_diff` never produces a `Report` where the same path
-    // appears in both `files`/`skipped` or both `tests`/`skipped` (see
-    // `TreeBuilder::insert_file`/`insert_test_file`/`insert_skipped`'s own
-    // doc comments), so `#[cfg(debug_assertions)]` keeps these panic-path
-    // tests out of release builds, matching the `debug_assert!`s themselves
-    // — they only guard a caller contract, not a condition `build_tree`
-    // needs to handle gracefully at runtime. `files`/`tests` overlapping on
-    // the same path, in contrast, is a *valid* `analyze_diff` output (a
-    // mixed file) — see `should_keep_real_symbols_when_file_is_also_in_tests`
-    // below, not a panic case.
-    // `build_tree` visits `report.files` before `report.skipped` (its own
-    // doc comment's discovery order), so this hits `insert_skipped`'s own
-    // assert, not `insert_file`'s — `insert_file` only guards against a
-    // path *already* marked skipped when files runs, which is not yet true
-    // here.
-    #[cfg(debug_assertions)]
-    #[test]
-    #[should_panic(expected = "report.skipped must not overlap report.files/report.tests")]
-    fn should_panic_when_the_same_path_appears_in_files_and_skipped() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![symbol("lib.rs::foo", "foo", SymbolKind::Function)],
-            }],
-            skipped: vec![SkippedFile {
-                path: "lib.rs".to_string(),
-                reason: rinkaku_core::render::SkipReason::Binary,
-            }],
-            ..empty_report()
-        };
-
-        build_tree(&report);
-    }
-
-    #[cfg(debug_assertions)]
-    #[test]
-    #[should_panic(expected = "report.skipped must not overlap report.files/report.tests")]
-    fn should_panic_when_the_same_path_appears_in_tests_and_skipped() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            tests: vec![TestFileSummary {
-                path: "lib.rs".to_string(),
-                symbol_count: 1,
-            }],
-            skipped: vec![SkippedFile {
-                path: "lib.rs".to_string(),
-                reason: rinkaku_core::render::SkipReason::Binary,
-            }],
-            ..empty_report()
-        };
-
-        build_tree(&report);
-    }
-
-    // Regression test (post-rebase integration check): `lib.rs` in both
-    // `report.files` (some real symbols) and `report.tests` (a test count
-    // for the rest) is exactly what `pipeline::partition_test_symbols`
-    // produces for a mixed file — e.g. a Rust file with production
-    // functions changed alongside its own `#[cfg(test)] mod tests` in the
-    // same diff (this crate's own `rinkaku-tui/src/app.rs` hit this in a
-    // live dogfood run). `build_tree` must keep both pieces of information
-    // on the one `TreeNode` rather than panicking or silently dropping
-    // either half.
-    #[test]
-    fn should_keep_real_symbols_when_file_is_also_in_tests() {
-        let report = Report {
-            origin: rinkaku_core::render::ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "lib.rs".to_string(),
-                symbols: vec![symbol("lib.rs::foo", "foo", SymbolKind::Function)],
-            }],
-            tests: vec![TestFileSummary {
-                path: "lib.rs".to_string(),
-                symbol_count: 3,
-            }],
-            ..empty_report()
-        };
-
-        let tree = build_tree(&report);
-
-        assert_eq!(1, tree.roots[0].children.len());
-        assert_eq!(Some(3), tree.roots[0].test_symbol_count);
-        assert_eq!(None, tree.roots[0].skip_reason);
-    }
-}
+#[path = "tree_tests/mod.rs"]
+mod tests;
